+++ app/views/main.scala.html
@@ -16,7 +16,7 @@
         <link rel="stylesheet" href="@routes.Assets.versioned("lib/font-awesome/css/all.css")">
         <link rel="shortcut icon" type="image/png" href="@routes.Assets.versioned("images/favicon.png")">
         <script src="@routes.Assets.versioned("lib/axios/dist/axios.min.js")" type="text/javascript"></script>
-        <script src="@routes.Assets.versioned("javascripts/main.js")" type="text/javascript"></script>
+        <script type="module" src="@routes.Assets.versioned("javascripts/main.js")" type="text/javascript"></script>
     </head>
     <body>
         <script src="@routes.Assets.versioned("lib/bootstrap/js/bootstrap.bundle.js")"></script>
