+++ app/views/posts/index.scala.html
@@ -56,9 +56,9 @@ UserOptRequestHeader,flash: Flash)
 				}
 			</div>
 			<section class="mb-5">
-				@form(routes.PostController.insert(),Symbol("name") -> "postForm") {
+				@form(routes.PostController.insert(),Symbol("id") -> "postForm", Symbol("name") -> "postForm") {
 					@CSRF.formField
-					@textarea(postForm("content"),Symbol("class")->"form-control", Symbol("placeholder")->"今日の豆知識をつぶやこう!")
+					@textarea(postForm("content"),Symbol("id")->"postContent",Symbol("class")->"form-control", Symbol("placeholder")->"今日の豆知識をつぶやこう!")
 					<div class="mt-1 text-end">
 						<span class="me-5">
 							<button class="btn btn-primary btn-sm" type="submit">投稿する</button>
@@ -122,5 +122,5 @@ UserOptRequestHeader,flash: Flash)
 			</section>
 		</section>
 	</div>
-	<script src="@routes.Assets.versioned("javascripts/posts/index.js")"></script>
+	<script type="module" src="@routes.Assets.versioned("javascripts/post_index.js")"></script>
 	}
