+++ public/javascripts/main.js
@@ -1,163 +1,2 @@
-// axiosインスタンス
-const instance = axios.create({
-  baseURL: "/",
-  timeout: 1000,
-  headers: { "Content-Type": "application/json" },
-});
-
-//サインインページにリダイレクト(認証)
-const needSignInFunction = (error) => {
-  if (error.response.status == 401) {
-    location.href = "/users/sign-in" + "?=needSignIn";
-  }
-};
-
-//like操作の抽象メソッド
-const likeManipulator = (url, likePostId, csrfToken, successFunction) => {
-  instance
-    .post(
-      url,
-      {
-        postId: likePostId,
-      },
-      {
-        headers: {
-          "Csrf-Token": csrfToken,
-        },
-      }
-    )
-    .then((response) => successFunction(response))
-    .catch((error) => {
-      needSignInFunction(error);
-    });
-};
-
-//likeのinsert操作抽象メソッド
-const insertLikeManipulator = (likePostId, csrfToken, insertSuccessFunction) =>
-  likeManipulator("/like/insert", likePostId, csrfToken, insertSuccessFunction);
-
-//likeのdelete操作抽象メソッド
-const deleteLikeManipulator = (likePostId, csrfToken, deleteSuccessFunction) =>
-  likeManipulator("/like/delete", likePostId, csrfToken, deleteSuccessFunction);
-
-//エラーメッセージを表示するノードを作成
-const createErrorMesssage = (errorMessage) => {
-  const errorMessageNode = document.createElement("div");
-  errorMessageNode.innerText = errorMessage;
-  errorMessageNode.style.color = "red";
-  errorMessageNode.style.paddingTop = "16px";
-  errorMessageNode.style.paddingLeft = "16px";
-  return errorMessageNode;
-};
-
-// エラーメッセージをDOMに追加
-const displayErrorMessage = (errorMessageNode, NodeThatHasError) => {
-  NodeThatHasError.parentNode.insertBefore(errorMessageNode, NodeThatHasError);
-};
-
-// エラーメッセージノードを削除
-const removeErrorMessage = (errorMessageNode) => {
-  if (errorMessageNode.previousElementSibling !== null) {
-    errorMessageNode.previousElementSibling.remove();
-  }
-};
-
-const formatDateUtil = (commentDate) => {
-  const month = commentDate.substring(5, 7);
-  const date = commentDate.substring(8, 10);
-  const hour = commentDate.substring(11, 13);
-  const minute = commentDate.substring(14, 16);
-  return `${month}/${date} ${hour}:${minute}`;
-};
-
-// コメントヘッダーの作成
-const createCommentCardHeader = (commentData) => {
-  const sectionHeader = document.createElement("section");
-  sectionHeader.classList.add("card-header");
-  sectionHeader.classList.add("bg-white");
-
-  //プロフィール画像img
-  const profileImg = document.createElement("img");
-  profileImg.src = `https://static.mameti-bbs.com/${commentData.userWhoCommented.profileImg}`;
-  profileImg.classList.add("profile-img-sm");
-  profileImg.setAttribute("alt", "プロフィール画像");
-
-  //ユーザ名span
-  const spanUserName = document.createElement("span");
-  spanUserName.classList.add("fs-5");
-  spanUserName.classList.add("fw-bold");
-  spanUserName.innerText = commentData.userWhoCommented.name;
-
-  //時刻span
-  const spanTime = document.createElement("span");
-  spanTime.classList.add("ps-3");
-  spanTime.classList.add("text-secondary");
-  spanTime.innerText = formatDateUtil(commentData.commentedAt);
-
-  //spanをsectionHeaderに追加
-  sectionHeader.append(profileImg);
-  sectionHeader.append(spanUserName);
-  sectionHeader.append(spanTime);
-  return sectionHeader;
-};
-
-//コメント いいね アイコンDivを作成
-const createIconsDiv = () => {
-  const iconsDiv = document.createElement("div");
-  iconsDiv.classList.add("px-3");
-  iconsDiv.classList.add("w-25");
-  iconsDiv.classList.add("d-flex");
-  iconsDiv.classList.add("justify-content-between");
-
-  //コメント
-  const comment = document.createElement("i");
-  comment.title = "コメントする";
-  comment.classList.add("far");
-  comment.classList.add("fa-comment");
-
-  //いいね
-  const like = document.createElement("i");
-  like.title = "いいね！";
-  like.classList.add("far");
-  like.classList.add("fa-heart");
-
-  //divにコメントといいねを追加
-  iconsDiv.append(comment);
-  iconsDiv.append(like);
-  return iconsDiv;
-};
-
-const createCommentCardBody = (commentData) => {
-  // カード内ボディ
-  const sectionBody = document.createElement("section");
-  sectionBody.classList.add("card-body");
-
-  // コメントコンテンツ
-  const contentDiv = document.createElement("div");
-  contentDiv.classList.add("card-text");
-  contentDiv.classList.add("p-3");
-  contentDiv.innerText = commentData.content;
-
-  const iconsDiv = createIconsDiv();
-
-  //sectionBodyにコンテンツとアイコンDivを追加
-  sectionBody.append(contentDiv);
-  sectionBody.append(iconsDiv);
-
-  return sectionBody;
-};
-
-//作成されたコメントをコメントリストに追加
-const appendCommentArticle = (parentNode, commentData) => {
-  const article = document.createElement("article");
-  article.classList.add("border");
-  article.classList.add("rounded");
-
-  const commentCardHeader = createCommentCardHeader(commentData);
-  const commentCardBody = createCommentCardBody(commentData);
-
-  article.appendChild(commentCardHeader);
-  article.appendChild(commentCardBody);
-
-  parentNode.prepend(article);
-};
+var M={exports:{}},se=function(e,t){return function(){for(var n=new Array(arguments.length),i=0;i<n.length;i++)n[i]=arguments[i];return e.apply(t,n)}},we=se,E=Object.prototype.toString;function H(r){return Array.isArray(r)}function I(r){return typeof r=="undefined"}function Se(r){return r!==null&&!I(r)&&r.constructor!==null&&!I(r.constructor)&&typeof r.constructor.isBuffer=="function"&&r.constructor.isBuffer(r)}function ie(r){return E.call(r)==="[object ArrayBuffer]"}function Ce(r){return E.call(r)==="[object FormData]"}function Re(r){var e;return typeof ArrayBuffer!="undefined"&&ArrayBuffer.isView?e=ArrayBuffer.isView(r):e=r&&r.buffer&&ie(r.buffer),e}function Oe(r){return typeof r=="string"}function xe(r){return typeof r=="number"}function oe(r){return r!==null&&typeof r=="object"}function $(r){if(E.call(r)!=="[object Object]")return!1;var e=Object.getPrototypeOf(r);return e===null||e===Object.prototype}function Ae(r){return E.call(r)==="[object Date]"}function Ne(r){return E.call(r)==="[object File]"}function $e(r){return E.call(r)==="[object Blob]"}function ue(r){return E.call(r)==="[object Function]"}function Te(r){return oe(r)&&ue(r.pipe)}function ge(r){return E.call(r)==="[object URLSearchParams]"}function Pe(r){return r.trim?r.trim():r.replace(/^\s+|\s+$/g,"")}function Ue(){return typeof navigator!="undefined"&&(navigator.product==="ReactNative"||navigator.product==="NativeScript"||navigator.product==="NS")?!1:typeof window!="undefined"&&typeof document!="undefined"}function J(r,e){if(!(r===null||typeof r=="undefined"))if(typeof r!="object"&&(r=[r]),H(r))for(var t=0,a=r.length;t<a;t++)e.call(null,r[t],t,r);else for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&e.call(null,r[n],n,r)}function F(){var r={};function e(n,i){$(r[i])&&$(n)?r[i]=F(r[i],n):$(n)?r[i]=F({},n):H(n)?r[i]=n.slice():r[i]=n}for(var t=0,a=arguments.length;t<a;t++)J(arguments[t],e);return r}function Le(r,e,t){return J(e,function(n,i){t&&typeof n=="function"?r[i]=we(n,t):r[i]=n}),r}function je(r){return r.charCodeAt(0)===65279&&(r=r.slice(1)),r}var h={isArray:H,isArrayBuffer:ie,isBuffer:Se,isFormData:Ce,isArrayBufferView:Re,isString:Oe,isNumber:xe,isObject:oe,isPlainObject:$,isUndefined:I,isDate:Ae,isFile:Ne,isBlob:$e,isFunction:ue,isStream:Te,isURLSearchParams:ge,isStandardBrowserEnv:Ue,forEach:J,merge:F,extend:Le,trim:Pe,stripBOM:je},C=h;function K(r){return encodeURIComponent(r).replace(/%3A/gi,":").replace(/%24/g,"$").replace(/%2C/gi,",").replace(/%20/g,"+").replace(/%5B/gi,"[").replace(/%5D/gi,"]")}var fe=function(e,t,a){if(!t)return e;var n;if(a)n=a(t);else if(C.isURLSearchParams(t))n=t.toString();else{var i=[];C.forEach(t,function(f,m){f===null||typeof f=="undefined"||(C.isArray(f)?m=m+"[]":f=[f],C.forEach(f,function(l){C.isDate(l)?l=l.toISOString():C.isObject(l)&&(l=JSON.stringify(l)),i.push(K(m)+"="+K(l))}))}),n=i.join("&")}if(n){var o=e.indexOf("#");o!==-1&&(e=e.slice(0,o)),e+=(e.indexOf("?")===-1?"?":"&")+n}return e},Be=h;function g(){this.handlers=[]}g.prototype.use=function(e,t,a){return this.handlers.push({fulfilled:e,rejected:t,synchronous:a?a.synchronous:!1,runWhen:a?a.runWhen:null}),this.handlers.length-1};g.prototype.eject=function(e){this.handlers[e]&&(this.handlers[e]=null)};g.prototype.forEach=function(e){Be.forEach(this.handlers,function(a){a!==null&&e(a)})};var De=g,qe=h,_e=function(e,t){qe.forEach(e,function(n,i){i!==t&&i.toUpperCase()===t.toUpperCase()&&(e[t]=n,delete e[i])})},le=function(e,t,a,n,i){return e.config=t,a&&(e.code=a),e.request=n,e.response=i,e.isAxiosError=!0,e.toJSON=function(){return{message:this.message,name:this.name,description:this.description,number:this.number,fileName:this.fileName,lineNumber:this.lineNumber,columnNumber:this.columnNumber,stack:this.stack,config:this.config,code:this.code,status:this.response&&this.response.status?this.response.status:null}},e},ke=le,ce=function(e,t,a,n,i){var o=new Error(e);return ke(o,t,a,n,i)},Ie=ce,Fe=function(e,t,a){var n=a.config.validateStatus;!a.status||!n||n(a.status)?e(a):t(Ie("Request failed with status code "+a.status,a.config,null,a.request,a))},A=h,Me=A.isStandardBrowserEnv()?function(){return{write:function(t,a,n,i,o,u){var f=[];f.push(t+"="+encodeURIComponent(a)),A.isNumber(n)&&f.push("expires="+new Date(n).toGMTString()),A.isString(i)&&f.push("path="+i),A.isString(o)&&f.push("domain="+o),u===!0&&f.push("secure"),document.cookie=f.join("; ")},read:function(t){var a=document.cookie.match(new RegExp("(^|;\\s*)("+t+")=([^;]*)"));return a?decodeURIComponent(a[3]):null},remove:function(t){this.write(t,"",Date.now()-864e5)}}}():function(){return{write:function(){},read:function(){return null},remove:function(){}}}(),He=function(e){return/^([a-z][a-z\d+\-.]*:)?\/\//i.test(e)},Je=function(e,t){return t?e.replace(/\/+$/,"")+"/"+t.replace(/^\/+/,""):e},ze=He,Ve=Je,We=function(e,t){return e&&!ze(t)?Ve(e,t):t},D=h,Xe=["age","authorization","content-length","content-type","etag","expires","from","host","if-modified-since","if-unmodified-since","last-modified","location","max-forwards","proxy-authorization","referer","retry-after","user-agent"],Ke=function(e){var t={},a,n,i;return e&&D.forEach(e.split(`
+`),function(u){if(i=u.indexOf(":"),a=D.trim(u.substr(0,i)).toLowerCase(),n=D.trim(u.substr(i+1)),a){if(t[a]&&Xe.indexOf(a)>=0)return;a==="set-cookie"?t[a]=(t[a]?t[a]:[]).concat([n]):t[a]=t[a]?t[a]+", "+n:n}}),t},G=h,Ge=G.isStandardBrowserEnv()?function(){var e=/(msie|trident)/i.test(navigator.userAgent),t=document.createElement("a"),a;function n(i){var o=i;return e&&(t.setAttribute("href",o),o=t.href),t.setAttribute("href",o),{href:t.href,protocol:t.protocol?t.protocol.replace(/:$/,""):"",host:t.host,search:t.search?t.search.replace(/^\?/,""):"",hash:t.hash?t.hash.replace(/^#/,""):"",hostname:t.hostname,port:t.port,pathname:t.pathname.charAt(0)==="/"?t.pathname:"/"+t.pathname}}return a=n(window.location.href),function(o){var u=G.isString(o)?n(o):o;return u.protocol===a.protocol&&u.host===a.host}}():function(){return function(){return!0}}();function z(r){this.message=r}z.prototype.toString=function(){return"Cancel"+(this.message?": "+this.message:"")};z.prototype.__CANCEL__=!0;var P=z,N=h,Ye=Fe,Qe=Me,Ze=fe,er=We,rr=Ke,tr=Ge,q=ce,nr=L,ar=P,Y=function(e){return new Promise(function(a,n){var i=e.data,o=e.headers,u=e.responseType,f;function m(){e.cancelToken&&e.cancelToken.unsubscribe(f),e.signal&&e.signal.removeEventListener("abort",f)}N.isFormData(i)&&delete o["Content-Type"];var s=new XMLHttpRequest;if(e.auth){var l=e.auth.username||"",y=e.auth.password?unescape(encodeURIComponent(e.auth.password)):"";o.Authorization="Basic "+btoa(l+":"+y)}var d=er(e.baseURL,e.url);s.open(e.method.toUpperCase(),Ze(d,e.params,e.paramsSerializer),!0),s.timeout=e.timeout;function W(){if(!!s){var v="getAllResponseHeaders"in s?rr(s.getAllResponseHeaders()):null,S=!u||u==="text"||u==="json"?s.responseText:s.response,w={data:S,status:s.status,statusText:s.statusText,headers:v,config:e,request:s};Ye(function(B){a(B),m()},function(B){n(B),m()},w),s=null}}if("onloadend"in s?s.onloadend=W:s.onreadystatechange=function(){!s||s.readyState!==4||s.status===0&&!(s.responseURL&&s.responseURL.indexOf("file:")===0)||setTimeout(W)},s.onabort=function(){!s||(n(q("Request aborted",e,"ECONNABORTED",s)),s=null)},s.onerror=function(){n(q("Network Error",e,null,s)),s=null},s.ontimeout=function(){var S=e.timeout?"timeout of "+e.timeout+"ms exceeded":"timeout exceeded",w=e.transitional||nr.transitional;e.timeoutErrorMessage&&(S=e.timeoutErrorMessage),n(q(S,e,w.clarifyTimeoutError?"ETIMEDOUT":"ECONNABORTED",s)),s=null},N.isStandardBrowserEnv()){var X=(e.withCredentials||tr(d))&&e.xsrfCookieName?Qe.read(e.xsrfCookieName):void 0;X&&(o[e.xsrfHeaderName]=X)}"setRequestHeader"in s&&N.forEach(o,function(S,w){typeof i=="undefined"&&w.toLowerCase()==="content-type"?delete o[w]:s.setRequestHeader(w,S)}),N.isUndefined(e.withCredentials)||(s.withCredentials=!!e.withCredentials),u&&u!=="json"&&(s.responseType=e.responseType),typeof e.onDownloadProgress=="function"&&s.addEventListener("progress",e.onDownloadProgress),typeof e.onUploadProgress=="function"&&s.upload&&s.upload.addEventListener("progress",e.onUploadProgress),(e.cancelToken||e.signal)&&(f=function(v){!s||(n(!v||v&&v.type?new ar("canceled"):v),s.abort(),s=null)},e.cancelToken&&e.cancelToken.subscribe(f),e.signal&&(e.signal.aborted?f():e.signal.addEventListener("abort",f))),i||(i=null),s.send(i)})},c=h,Q=_e,sr=le,ir={"Content-Type":"application/x-www-form-urlencoded"};function Z(r,e){!c.isUndefined(r)&&c.isUndefined(r["Content-Type"])&&(r["Content-Type"]=e)}function or(){var r;return(typeof XMLHttpRequest!="undefined"||typeof process!="undefined"&&Object.prototype.toString.call(process)==="[object process]")&&(r=Y),r}function ur(r,e,t){if(c.isString(r))try{return(e||JSON.parse)(r),c.trim(r)}catch(a){if(a.name!=="SyntaxError")throw a}return(t||JSON.stringify)(r)}var U={transitional:{silentJSONParsing:!0,forcedJSONParsing:!0,clarifyTimeoutError:!1},adapter:or(),transformRequest:[function(e,t){return Q(t,"Accept"),Q(t,"Content-Type"),c.isFormData(e)||c.isArrayBuffer(e)||c.isBuffer(e)||c.isStream(e)||c.isFile(e)||c.isBlob(e)?e:c.isArrayBufferView(e)?e.buffer:c.isURLSearchParams(e)?(Z(t,"application/x-www-form-urlencoded;charset=utf-8"),e.toString()):c.isObject(e)||t&&t["Content-Type"]==="application/json"?(Z(t,"application/json"),ur(e)):e}],transformResponse:[function(e){var t=this.transitional||U.transitional,a=t&&t.silentJSONParsing,n=t&&t.forcedJSONParsing,i=!a&&this.responseType==="json";if(i||n&&c.isString(e)&&e.length)try{return JSON.parse(e)}catch(o){if(i)throw o.name==="SyntaxError"?sr(o,this,"E_JSON_PARSE"):o}return e}],timeout:0,xsrfCookieName:"XSRF-TOKEN",xsrfHeaderName:"X-XSRF-TOKEN",maxContentLength:-1,maxBodyLength:-1,validateStatus:function(e){return e>=200&&e<300},headers:{common:{Accept:"application/json, text/plain, */*"}}};c.forEach(["delete","get","head"],function(e){U.headers[e]={}});c.forEach(["post","put","patch"],function(e){U.headers[e]=c.merge(ir)});var L=U,fr=h,lr=L,cr=function(e,t,a){var n=this||lr;return fr.forEach(a,function(o){e=o.call(n,e,t)}),e},de=function(e){return!!(e&&e.__CANCEL__)},ee=h,_=cr,dr=de,hr=L,pr=P;function k(r){if(r.cancelToken&&r.cancelToken.throwIfRequested(),r.signal&&r.signal.aborted)throw new pr("canceled")}var mr=function(e){k(e),e.headers=e.headers||{},e.data=_.call(e,e.data,e.headers,e.transformRequest),e.headers=ee.merge(e.headers.common||{},e.headers[e.method]||{},e.headers),ee.forEach(["delete","get","head","post","put","patch","common"],function(n){delete e.headers[n]});var t=e.adapter||hr.adapter;return t(e).then(function(n){return k(e),n.data=_.call(e,n.data,n.headers,e.transformResponse),n},function(n){return dr(n)||(k(e),n&&n.response&&(n.response.data=_.call(e,n.response.data,n.response.headers,e.transformResponse))),Promise.reject(n)})},p=h,he=function(e,t){t=t||{};var a={};function n(s,l){return p.isPlainObject(s)&&p.isPlainObject(l)?p.merge(s,l):p.isPlainObject(l)?p.merge({},l):p.isArray(l)?l.slice():l}function i(s){if(p.isUndefined(t[s])){if(!p.isUndefined(e[s]))return n(void 0,e[s])}else return n(e[s],t[s])}function o(s){if(!p.isUndefined(t[s]))return n(void 0,t[s])}function u(s){if(p.isUndefined(t[s])){if(!p.isUndefined(e[s]))return n(void 0,e[s])}else return n(void 0,t[s])}function f(s){if(s in t)return n(e[s],t[s]);if(s in e)return n(void 0,e[s])}var m={url:o,method:o,data:o,baseURL:u,transformRequest:u,transformResponse:u,paramsSerializer:u,timeout:u,timeoutMessage:u,withCredentials:u,adapter:u,responseType:u,xsrfCookieName:u,xsrfHeaderName:u,onUploadProgress:u,onDownloadProgress:u,decompress:u,maxContentLength:u,maxBodyLength:u,transport:u,httpAgent:u,httpsAgent:u,cancelToken:u,socketPath:u,responseEncoding:u,validateStatus:f};return p.forEach(Object.keys(e).concat(Object.keys(t)),function(l){var y=m[l]||i,d=y(l);p.isUndefined(d)&&y!==f||(a[l]=d)}),a},pe={version:"0.26.0"},vr=pe.version,V={};["object","boolean","number","function","string","symbol"].forEach(function(r,e){V[r]=function(a){return typeof a===r||"a"+(e<1?"n ":" ")+r}});var re={};V.transitional=function(e,t,a){function n(i,o){return"[Axios v"+vr+"] Transitional option '"+i+"'"+o+(a?". "+a:"")}return function(i,o,u){if(e===!1)throw new Error(n(o," has been removed"+(t?" in "+t:"")));return t&&!re[o]&&(re[o]=!0,console.warn(n(o," has been deprecated since v"+t+" and will be removed in the near future"))),e?e(i,o,u):!0}};function yr(r,e,t){if(typeof r!="object")throw new TypeError("options must be an object");for(var a=Object.keys(r),n=a.length;n-- >0;){var i=a[n],o=e[i];if(o){var u=r[i],f=u===void 0||o(u,i,r);if(f!==!0)throw new TypeError("option "+i+" must be "+f);continue}if(t!==!0)throw Error("Unknown option "+i)}}var br={assertOptions:yr,validators:V},me=h,Er=fe,te=De,ne=mr,j=he,ve=br,R=ve.validators;function x(r){this.defaults=r,this.interceptors={request:new te,response:new te}}x.prototype.request=function(e,t){typeof e=="string"?(t=t||{},t.url=e):t=e||{},t=j(this.defaults,t),t.method?t.method=t.method.toLowerCase():this.defaults.method?t.method=this.defaults.method.toLowerCase():t.method="get";var a=t.transitional;a!==void 0&&ve.assertOptions(a,{silentJSONParsing:R.transitional(R.boolean),forcedJSONParsing:R.transitional(R.boolean),clarifyTimeoutError:R.transitional(R.boolean)},!1);var n=[],i=!0;this.interceptors.request.forEach(function(d){typeof d.runWhen=="function"&&d.runWhen(t)===!1||(i=i&&d.synchronous,n.unshift(d.fulfilled,d.rejected))});var o=[];this.interceptors.response.forEach(function(d){o.push(d.fulfilled,d.rejected)});var u;if(!i){var f=[ne,void 0];for(Array.prototype.unshift.apply(f,n),f=f.concat(o),u=Promise.resolve(t);f.length;)u=u.then(f.shift(),f.shift());return u}for(var m=t;n.length;){var s=n.shift(),l=n.shift();try{m=s(m)}catch(y){l(y);break}}try{u=ne(m)}catch(y){return Promise.reject(y)}for(;o.length;)u=u.then(o.shift(),o.shift());return u};x.prototype.getUri=function(e){return e=j(this.defaults,e),Er(e.url,e.params,e.paramsSerializer).replace(/^\?/,"")};me.forEach(["delete","get","head","options"],function(e){x.prototype[e]=function(t,a){return this.request(j(a||{},{method:e,url:t,data:(a||{}).data}))}});me.forEach(["post","put","patch"],function(e){x.prototype[e]=function(t,a,n){return this.request(j(n||{},{method:e,url:t,data:a}))}});var wr=x,Sr=P;function O(r){if(typeof r!="function")throw new TypeError("executor must be a function.");var e;this.promise=new Promise(function(n){e=n});var t=this;this.promise.then(function(a){if(!!t._listeners){var n,i=t._listeners.length;for(n=0;n<i;n++)t._listeners[n](a);t._listeners=null}}),this.promise.then=function(a){var n,i=new Promise(function(o){t.subscribe(o),n=o}).then(a);return i.cancel=function(){t.unsubscribe(n)},i},r(function(n){t.reason||(t.reason=new Sr(n),e(t.reason))})}O.prototype.throwIfRequested=function(){if(this.reason)throw this.reason};O.prototype.subscribe=function(e){if(this.reason){e(this.reason);return}this._listeners?this._listeners.push(e):this._listeners=[e]};O.prototype.unsubscribe=function(e){if(!!this._listeners){var t=this._listeners.indexOf(e);t!==-1&&this._listeners.splice(t,1)}};O.source=function(){var e,t=new O(function(n){e=n});return{token:t,cancel:e}};var Cr=O,Rr=function(e){return function(a){return e.apply(null,a)}},Or=h,xr=function(e){return Or.isObject(e)&&e.isAxiosError===!0},ae=h,Ar=se,T=wr,Nr=he,$r=L;function ye(r){var e=new T(r),t=Ar(T.prototype.request,e);return ae.extend(t,T.prototype,e),ae.extend(t,e),t.create=function(n){return ye(Nr(r,n))},t}var b=ye($r);b.Axios=T;b.Cancel=P;b.CancelToken=Cr;b.isCancel=de;b.VERSION=pe.version;b.all=function(e){return Promise.all(e)};b.spread=Rr;b.isAxiosError=xr;M.exports=b;M.exports.default=b;var Tr=M.exports;const gr=Tr.create({baseURL:"/",timeout:1e3,headers:{"Content-Type":"application/json"}}),Pr=r=>{r.response.status==401&&(location.href="/users/sign-in?=needSignIn")},be=(r,e,t,a)=>{gr.post(r,{postId:e},{headers:{"Csrf-Token":t}}).then(n=>a(n)).catch(n=>{Pr(n)})},Ur=(r,e,t)=>be("/like/insert",r,e,t),Lr=(r,e,t)=>be("/like/delete",r,e,t);export{Ur as a,Lr as d,gr as i};
