+++ public/javascripts/post_index.js
@@ -1,0 +1,5 @@
+import{i as Fi,a as Gi,d as Xi}from"./main.js";var $="top",V="bottom",H="right",I="left",Ie="auto",Mt=[$,V,H,I],yt="start",Lt="end",Nr="clippingParents",uo="viewport",Tt="popper",$r="reference",eo=Mt.reduce(function(o,t){return o.concat([t+"-"+yt,t+"-"+Lt])},[]),bo=[].concat(Mt,[Ie]).reduce(function(o,t){return o.concat([t,t+"-"+yt,t+"-"+Lt])},[]),Ir="beforeRead",zr="read",Mr="afterRead",Rr="beforeMain",Pr="main",jr="afterMain",Vr="beforeWrite",Hr="write",Br="afterWrite",Wr=[Ir,zr,Mr,Rr,Pr,jr,Vr,Hr,Br];function Z(o){return o?(o.nodeName||"").toLowerCase():null}function q(o){if(o==null)return window;if(o.toString()!=="[object Window]"){var t=o.ownerDocument;return t&&t.defaultView||window}return o}function Ct(o){var t=q(o).Element;return o instanceof t||o instanceof Element}function K(o){var t=q(o).HTMLElement;return o instanceof t||o instanceof HTMLElement}function ho(o){if(typeof ShadowRoot=="undefined")return!1;var t=q(o).ShadowRoot;return o instanceof t||o instanceof ShadowRoot}function qi(o){var t=o.state;Object.keys(t.elements).forEach(function(e){var r=t.styles[e]||{},i=t.attributes[e]||{},n=t.elements[e];!K(n)||!Z(n)||(Object.assign(n.style,r),Object.keys(i).forEach(function(a){var l=i[a];l===!1?n.removeAttribute(a):n.setAttribute(a,l===!0?"":l)}))})}function Qi(o){var t=o.state,e={popper:{position:t.options.strategy,left:"0",top:"0",margin:"0"},arrow:{position:"absolute"},reference:{}};return Object.assign(t.elements.popper.style,e.popper),t.styles=e,t.elements.arrow&&Object.assign(t.elements.arrow.style,e.arrow),function(){Object.keys(t.elements).forEach(function(r){var i=t.elements[r],n=t.attributes[r]||{},a=Object.keys(t.styles.hasOwnProperty(r)?t.styles[r]:e[r]),l=a.reduce(function(d,c){return d[c]="",d},{});!K(i)||!Z(i)||(Object.assign(i.style,l),Object.keys(n).forEach(function(d){i.removeAttribute(d)}))})}}var xo={name:"applyStyles",enabled:!0,phase:"write",fn:qi,effect:Qi,requires:["computeStyles"]};function J(o){return o.split("-")[0]}var vt=Math.max,Oe=Math.min,Dt=Math.round;function Nt(o,t){t===void 0&&(t=!1);var e=o.getBoundingClientRect(),r=1,i=1;if(K(o)&&t){var n=o.offsetHeight,a=o.offsetWidth;a>0&&(r=Dt(e.width)/a||1),n>0&&(i=Dt(e.height)/n||1)}return{width:e.width/r,height:e.height/i,top:e.top/i,right:e.right/r,bottom:e.bottom/i,left:e.left/r,x:e.left/r,y:e.top/i}}function vo(o){var t=Nt(o),e=o.offsetWidth,r=o.offsetHeight;return Math.abs(t.width-e)<=1&&(e=t.width),Math.abs(t.height-r)<=1&&(r=t.height),{x:o.offsetLeft,y:o.offsetTop,width:e,height:r}}function Kr(o,t){var e=t.getRootNode&&t.getRootNode();if(o.contains(t))return!0;if(e&&ho(e)){var r=t;do{if(r&&o.isSameNode(r))return!0;r=r.parentNode||r.host}while(r)}return!1}function et(o){return q(o).getComputedStyle(o)}function Ji(o){return["table","td","th"].indexOf(Z(o))>=0}function st(o){return((Ct(o)?o.ownerDocument:o.document)||window.document).documentElement}function ze(o){return Z(o)==="html"?o:o.assignedSlot||o.parentNode||(ho(o)?o.host:null)||st(o)}function Ko(o){return!K(o)||et(o).position==="fixed"?null:o.offsetParent}function Zi(o){var t=navigator.userAgent.toLowerCase().indexOf("firefox")!==-1,e=navigator.userAgent.indexOf("Trident")!==-1;if(e&&K(o)){var r=et(o);if(r.position==="fixed")return null}var i=ze(o);for(ho(i)&&(i=i.host);K(i)&&["html","body"].indexOf(Z(i))<0;){var n=et(i);if(n.transform!=="none"||n.perspective!=="none"||n.contain==="paint"||["transform","perspective"].indexOf(n.willChange)!==-1||t&&n.willChange==="filter"||t&&n.filter&&n.filter!=="none")return i;i=i.parentNode}return null}function ne(o){for(var t=q(o),e=Ko(o);e&&Ji(e)&&et(e).position==="static";)e=Ko(e);return e&&(Z(e)==="html"||Z(e)==="body"&&et(e).position==="static")?t:e||Zi(o)||t}function wo(o){return["top","bottom"].indexOf(o)>=0?"x":"y"}function Jt(o,t,e){return vt(o,Oe(t,e))}function tn(o,t,e){var r=Jt(o,t,e);return r>e?e:r}function Yr(){return{top:0,right:0,bottom:0,left:0}}function Ur(o){return Object.assign({},Yr(),o)}function Fr(o,t){return t.reduce(function(e,r){return e[r]=o,e},{})}var en=function(t,e){return t=typeof t=="function"?t(Object.assign({},e.rects,{placement:e.placement})):t,Ur(typeof t!="number"?t:Fr(t,Mt))};function on(o){var t,e=o.state,r=o.name,i=o.options,n=e.elements.arrow,a=e.modifiersData.popperOffsets,l=J(e.placement),d=wo(l),c=[I,H].indexOf(l)>=0,m=c?"height":"width";if(!(!n||!a)){var p=en(i.padding,e),b=vo(n),g=d==="y"?$:I,x=d==="y"?V:H,u=e.rects.reference[m]+e.rects.reference[d]-a[d]-e.rects.popper[m],v=a[d]-e.rects.reference[d],E=ne(n),T=E?d==="y"?E.clientHeight||0:E.clientWidth||0:0,A=u/2-v/2,h=p[g],_=T-b[m]-p[x],w=T/2-b[m]/2+A,k=Jt(h,w,_),S=d;e.modifiersData[r]=(t={},t[S]=k,t.centerOffset=k-w,t)}}function rn(o){var t=o.state,e=o.options,r=e.element,i=r===void 0?"[data-popper-arrow]":r;i!=null&&(typeof i=="string"&&(i=t.elements.popper.querySelector(i),!i)||!Kr(t.elements.popper,i)||(t.elements.arrow=i))}var Gr={name:"arrow",enabled:!0,phase:"main",fn:on,effect:rn,requires:["popperOffsets"],requiresIfExists:["preventOverflow"]};function $t(o){return o.split("-")[1]}var nn={top:"auto",right:"auto",bottom:"auto",left:"auto"};function an(o){var t=o.x,e=o.y,r=window,i=r.devicePixelRatio||1;return{x:Dt(t*i)/i||0,y:Dt(e*i)/i||0}}function Yo(o){var t,e=o.popper,r=o.popperRect,i=o.placement,n=o.variation,a=o.offsets,l=o.position,d=o.gpuAcceleration,c=o.adaptive,m=o.roundOffsets,p=o.isFixed,b=a.x,g=b===void 0?0:b,x=a.y,u=x===void 0?0:x,v=typeof m=="function"?m({x:g,y:u}):{x:g,y:u};g=v.x,u=v.y;var E=a.hasOwnProperty("x"),T=a.hasOwnProperty("y"),A=I,h=$,_=window;if(c){var w=ne(e),k="clientHeight",S="clientWidth";if(w===q(e)&&(w=st(e),et(w).position!=="static"&&l==="absolute"&&(k="scrollHeight",S="scrollWidth")),w=w,i===$||(i===I||i===H)&&n===Lt){h=V;var D=p&&w===_&&_.visualViewport?_.visualViewport.height:w[k];u-=D-r.height,u*=d?1:-1}if(i===I||(i===$||i===V)&&n===Lt){A=H;var L=p&&w===_&&_.visualViewport?_.visualViewport.width:w[S];g-=L-r.width,g*=d?1:-1}}var C=Object.assign({position:l},c&&nn),U=m===!0?an({x:g,y:u}):{x:g,y:u};if(g=U.x,u=U.y,d){var R;return Object.assign({},C,(R={},R[h]=T?"0":"",R[A]=E?"0":"",R.transform=(_.devicePixelRatio||1)<=1?"translate("+g+"px, "+u+"px)":"translate3d("+g+"px, "+u+"px, 0)",R))}return Object.assign({},C,(t={},t[h]=T?u+"px":"",t[A]=E?g+"px":"",t.transform="",t))}function ln(o){var t=o.state,e=o.options,r=e.gpuAcceleration,i=r===void 0?!0:r,n=e.adaptive,a=n===void 0?!0:n,l=e.roundOffsets,d=l===void 0?!0:l,c={placement:J(t.placement),variation:$t(t.placement),popper:t.elements.popper,popperRect:t.rects.popper,gpuAcceleration:i,isFixed:t.options.strategy==="fixed"};t.modifiersData.popperOffsets!=null&&(t.styles.popper=Object.assign({},t.styles.popper,Yo(Object.assign({},c,{offsets:t.modifiersData.popperOffsets,position:t.options.strategy,adaptive:a,roundOffsets:d})))),t.modifiersData.arrow!=null&&(t.styles.arrow=Object.assign({},t.styles.arrow,Yo(Object.assign({},c,{offsets:t.modifiersData.arrow,position:"absolute",adaptive:!1,roundOffsets:d})))),t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-placement":t.placement})}var yo={name:"computeStyles",enabled:!0,phase:"beforeWrite",fn:ln,data:{}},fe={passive:!0};function sn(o){var t=o.state,e=o.instance,r=o.options,i=r.scroll,n=i===void 0?!0:i,a=r.resize,l=a===void 0?!0:a,d=q(t.elements.popper),c=[].concat(t.scrollParents.reference,t.scrollParents.popper);return n&&c.forEach(function(m){m.addEventListener("scroll",e.update,fe)}),l&&d.addEventListener("resize",e.update,fe),function(){n&&c.forEach(function(m){m.removeEventListener("scroll",e.update,fe)}),l&&d.removeEventListener("resize",e.update,fe)}}var _o={name:"eventListeners",enabled:!0,phase:"write",fn:function(){},effect:sn,data:{}},dn={left:"right",right:"left",bottom:"top",top:"bottom"};function _e(o){return o.replace(/left|right|bottom|top/g,function(t){return dn[t]})}var mn={start:"end",end:"start"};function Uo(o){return o.replace(/start|end/g,function(t){return mn[t]})}function ko(o){var t=q(o),e=t.pageXOffset,r=t.pageYOffset;return{scrollLeft:e,scrollTop:r}}function Eo(o){return Nt(st(o)).left+ko(o).scrollLeft}function cn(o){var t=q(o),e=st(o),r=t.visualViewport,i=e.clientWidth,n=e.clientHeight,a=0,l=0;return r&&(i=r.width,n=r.height,/^((?!chrome|android).)*safari/i.test(navigator.userAgent)||(a=r.offsetLeft,l=r.offsetTop)),{width:i,height:n,x:a+Eo(o),y:l}}function pn(o){var t,e=st(o),r=ko(o),i=(t=o.ownerDocument)==null?void 0:t.body,n=vt(e.scrollWidth,e.clientWidth,i?i.scrollWidth:0,i?i.clientWidth:0),a=vt(e.scrollHeight,e.clientHeight,i?i.scrollHeight:0,i?i.clientHeight:0),l=-r.scrollLeft+Eo(o),d=-r.scrollTop;return et(i||e).direction==="rtl"&&(l+=vt(e.clientWidth,i?i.clientWidth:0)-n),{width:n,height:a,x:l,y:d}}function To(o){var t=et(o),e=t.overflow,r=t.overflowX,i=t.overflowY;return/auto|scroll|overlay|hidden/.test(e+i+r)}function Xr(o){return["html","body","#document"].indexOf(Z(o))>=0?o.ownerDocument.body:K(o)&&To(o)?o:Xr(ze(o))}function Zt(o,t){var e;t===void 0&&(t=[]);var r=Xr(o),i=r===((e=o.ownerDocument)==null?void 0:e.body),n=q(r),a=i?[n].concat(n.visualViewport||[],To(r)?r:[]):r,l=t.concat(a);return i?l:l.concat(Zt(ze(a)))}function oo(o){return Object.assign({},o,{left:o.x,top:o.y,right:o.x+o.width,bottom:o.y+o.height})}function gn(o){var t=Nt(o);return t.top=t.top+o.clientTop,t.left=t.left+o.clientLeft,t.bottom=t.top+o.clientHeight,t.right=t.left+o.clientWidth,t.width=o.clientWidth,t.height=o.clientHeight,t.x=t.left,t.y=t.top,t}function Fo(o,t){return t===uo?oo(cn(o)):Ct(t)?gn(t):oo(pn(st(o)))}function fn(o){var t=Zt(ze(o)),e=["absolute","fixed"].indexOf(et(o).position)>=0,r=e&&K(o)?ne(o):o;return Ct(r)?t.filter(function(i){return Ct(i)&&Kr(i,r)&&Z(i)!=="body"}):[]}function un(o,t,e){var r=t==="clippingParents"?fn(o):[].concat(t),i=[].concat(r,[e]),n=i[0],a=i.reduce(function(l,d){var c=Fo(o,d);return l.top=vt(c.top,l.top),l.right=Oe(c.right,l.right),l.bottom=Oe(c.bottom,l.bottom),l.left=vt(c.left,l.left),l},Fo(o,n));return a.width=a.right-a.left,a.height=a.bottom-a.top,a.x=a.left,a.y=a.top,a}function qr(o){var t=o.reference,e=o.element,r=o.placement,i=r?J(r):null,n=r?$t(r):null,a=t.x+t.width/2-e.width/2,l=t.y+t.height/2-e.height/2,d;switch(i){case $:d={x:a,y:t.y-e.height};break;case V:d={x:a,y:t.y+t.height};break;case H:d={x:t.x+t.width,y:l};break;case I:d={x:t.x-e.width,y:l};break;default:d={x:t.x,y:t.y}}var c=i?wo(i):null;if(c!=null){var m=c==="y"?"height":"width";switch(n){case yt:d[c]=d[c]-(t[m]/2-e[m]/2);break;case Lt:d[c]=d[c]+(t[m]/2-e[m]/2);break}}return d}function It(o,t){t===void 0&&(t={});var e=t,r=e.placement,i=r===void 0?o.placement:r,n=e.boundary,a=n===void 0?Nr:n,l=e.rootBoundary,d=l===void 0?uo:l,c=e.elementContext,m=c===void 0?Tt:c,p=e.altBoundary,b=p===void 0?!1:p,g=e.padding,x=g===void 0?0:g,u=Ur(typeof x!="number"?x:Fr(x,Mt)),v=m===Tt?$r:Tt,E=o.rects.popper,T=o.elements[b?v:m],A=un(Ct(T)?T:T.contextElement||st(o.elements.popper),a,d),h=Nt(o.elements.reference),_=qr({reference:h,element:E,strategy:"absolute",placement:i}),w=oo(Object.assign({},E,_)),k=m===Tt?w:h,S={top:A.top-k.top+u.top,bottom:k.bottom-A.bottom+u.bottom,left:A.left-k.left+u.left,right:k.right-A.right+u.right},D=o.modifiersData.offset;if(m===Tt&&D){var L=D[i];Object.keys(S).forEach(function(C){var U=[H,V].indexOf(C)>=0?1:-1,R=[$,V].indexOf(C)>=0?"y":"x";S[C]+=L[R]*U})}return S}function bn(o,t){t===void 0&&(t={});var e=t,r=e.placement,i=e.boundary,n=e.rootBoundary,a=e.padding,l=e.flipVariations,d=e.allowedAutoPlacements,c=d===void 0?bo:d,m=$t(r),p=m?l?eo:eo.filter(function(x){return $t(x)===m}):Mt,b=p.filter(function(x){return c.indexOf(x)>=0});b.length===0&&(b=p);var g=b.reduce(function(x,u){return x[u]=It(o,{placement:u,boundary:i,rootBoundary:n,padding:a})[J(u)],x},{});return Object.keys(g).sort(function(x,u){return g[x]-g[u]})}function hn(o){if(J(o)===Ie)return[];var t=_e(o);return[Uo(o),t,Uo(t)]}function xn(o){var t=o.state,e=o.options,r=o.name;if(!t.modifiersData[r]._skip){for(var i=e.mainAxis,n=i===void 0?!0:i,a=e.altAxis,l=a===void 0?!0:a,d=e.fallbackPlacements,c=e.padding,m=e.boundary,p=e.rootBoundary,b=e.altBoundary,g=e.flipVariations,x=g===void 0?!0:g,u=e.allowedAutoPlacements,v=t.options.placement,E=J(v),T=E===v,A=d||(T||!x?[_e(v)]:hn(v)),h=[v].concat(A).reduce(function(Et,it){return Et.concat(J(it)===Ie?bn(t,{placement:it,boundary:m,rootBoundary:p,padding:c,flipVariations:x,allowedAutoPlacements:u}):it)},[]),_=t.rects.reference,w=t.rects.popper,k=new Map,S=!0,D=h[0],L=0;L<h.length;L++){var C=h[L],U=J(C),R=$t(C)===yt,jt=[$,V].indexOf(U)>=0,Vt=jt?"width":"height",P=It(t,{placement:C,boundary:m,rootBoundary:p,altBoundary:b,padding:c}),F=jt?R?H:I:R?V:$;_[Vt]>w[Vt]&&(F=_e(F));var de=_e(F),mt=[];if(n&&mt.push(P[U]<=0),l&&mt.push(P[F]<=0,P[de]<=0),mt.every(function(Et){return Et})){D=C,S=!1;break}k.set(C,mt)}if(S)for(var me=x?3:1,He=function(it){var Bt=h.find(function(pe){var ct=k.get(pe);if(ct)return ct.slice(0,it).every(function(Be){return Be})});if(Bt)return D=Bt,"break"},Ht=me;Ht>0;Ht--){var ce=He(Ht);if(ce==="break")break}t.placement!==D&&(t.modifiersData[r]._skip=!0,t.placement=D,t.reset=!0)}}var Qr={name:"flip",enabled:!0,phase:"main",fn:xn,requiresIfExists:["offset"],data:{_skip:!1}};function Go(o,t,e){return e===void 0&&(e={x:0,y:0}),{top:o.top-t.height-e.y,right:o.right-t.width+e.x,bottom:o.bottom-t.height+e.y,left:o.left-t.width-e.x}}function Xo(o){return[$,H,V,I].some(function(t){return o[t]>=0})}function vn(o){var t=o.state,e=o.name,r=t.rects.reference,i=t.rects.popper,n=t.modifiersData.preventOverflow,a=It(t,{elementContext:"reference"}),l=It(t,{altBoundary:!0}),d=Go(a,r),c=Go(l,i,n),m=Xo(d),p=Xo(c);t.modifiersData[e]={referenceClippingOffsets:d,popperEscapeOffsets:c,isReferenceHidden:m,hasPopperEscaped:p},t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-reference-hidden":m,"data-popper-escaped":p})}var Jr={name:"hide",enabled:!0,phase:"main",requiresIfExists:["preventOverflow"],fn:vn};function wn(o,t,e){var r=J(o),i=[I,$].indexOf(r)>=0?-1:1,n=typeof e=="function"?e(Object.assign({},t,{placement:o})):e,a=n[0],l=n[1];return a=a||0,l=(l||0)*i,[I,H].indexOf(r)>=0?{x:l,y:a}:{x:a,y:l}}function yn(o){var t=o.state,e=o.options,r=o.name,i=e.offset,n=i===void 0?[0,0]:i,a=bo.reduce(function(m,p){return m[p]=wn(p,t.rects,n),m},{}),l=a[t.placement],d=l.x,c=l.y;t.modifiersData.popperOffsets!=null&&(t.modifiersData.popperOffsets.x+=d,t.modifiersData.popperOffsets.y+=c),t.modifiersData[r]=a}var Zr={name:"offset",enabled:!0,phase:"main",requires:["popperOffsets"],fn:yn};function _n(o){var t=o.state,e=o.name;t.modifiersData[e]=qr({reference:t.rects.reference,element:t.rects.popper,strategy:"absolute",placement:t.placement})}var Ao={name:"popperOffsets",enabled:!0,phase:"read",fn:_n,data:{}};function kn(o){return o==="x"?"y":"x"}function En(o){var t=o.state,e=o.options,r=o.name,i=e.mainAxis,n=i===void 0?!0:i,a=e.altAxis,l=a===void 0?!1:a,d=e.boundary,c=e.rootBoundary,m=e.altBoundary,p=e.padding,b=e.tether,g=b===void 0?!0:b,x=e.tetherOffset,u=x===void 0?0:x,v=It(t,{boundary:d,rootBoundary:c,padding:p,altBoundary:m}),E=J(t.placement),T=$t(t.placement),A=!T,h=wo(E),_=kn(h),w=t.modifiersData.popperOffsets,k=t.rects.reference,S=t.rects.popper,D=typeof u=="function"?u(Object.assign({},t.rects,{placement:t.placement})):u,L=typeof D=="number"?{mainAxis:D,altAxis:D}:Object.assign({mainAxis:0,altAxis:0},D),C=t.modifiersData.offset?t.modifiersData.offset[t.placement]:null,U={x:0,y:0};if(!!w){if(n){var R,jt=h==="y"?$:I,Vt=h==="y"?V:H,P=h==="y"?"height":"width",F=w[h],de=F+v[jt],mt=F-v[Vt],me=g?-S[P]/2:0,He=T===yt?k[P]:S[P],Ht=T===yt?-S[P]:-k[P],ce=t.elements.arrow,Et=g&&ce?vo(ce):{width:0,height:0},it=t.modifiersData["arrow#persistent"]?t.modifiersData["arrow#persistent"].padding:Yr(),Bt=it[jt],pe=it[Vt],ct=Jt(0,k[P],Et[P]),Be=A?k[P]/2-me-ct-Bt-L.mainAxis:He-ct-Bt-L.mainAxis,Hi=A?-k[P]/2+me+ct+pe+L.mainAxis:Ht+ct+pe+L.mainAxis,We=t.elements.arrow&&ne(t.elements.arrow),Bi=We?h==="y"?We.clientTop||0:We.clientLeft||0:0,zo=(R=C==null?void 0:C[h])!=null?R:0,Wi=F+Be-zo-Bi,Ki=F+Hi-zo,Mo=Jt(g?Oe(de,Wi):de,F,g?vt(mt,Ki):mt);w[h]=Mo,U[h]=Mo-F}if(l){var Ro,Yi=h==="x"?$:I,Ui=h==="x"?V:H,pt=w[_],ge=_==="y"?"height":"width",Po=pt+v[Yi],jo=pt-v[Ui],Ke=[$,I].indexOf(E)!==-1,Vo=(Ro=C==null?void 0:C[_])!=null?Ro:0,Ho=Ke?Po:pt-k[ge]-S[ge]-Vo+L.altAxis,Bo=Ke?pt+k[ge]+S[ge]-Vo-L.altAxis:jo,Wo=g&&Ke?tn(Ho,pt,Bo):Jt(g?Ho:Po,pt,g?Bo:jo);w[_]=Wo,U[_]=Wo-pt}t.modifiersData[r]=U}}var ti={name:"preventOverflow",enabled:!0,phase:"main",fn:En,requiresIfExists:["offset"]};function Tn(o){return{scrollLeft:o.scrollLeft,scrollTop:o.scrollTop}}function An(o){return o===q(o)||!K(o)?ko(o):Tn(o)}function Sn(o){var t=o.getBoundingClientRect(),e=Dt(t.width)/o.offsetWidth||1,r=Dt(t.height)/o.offsetHeight||1;return e!==1||r!==1}function On(o,t,e){e===void 0&&(e=!1);var r=K(t),i=K(t)&&Sn(t),n=st(t),a=Nt(o,i),l={scrollLeft:0,scrollTop:0},d={x:0,y:0};return(r||!r&&!e)&&((Z(t)!=="body"||To(n))&&(l=An(t)),K(t)?(d=Nt(t,!0),d.x+=t.clientLeft,d.y+=t.clientTop):n&&(d.x=Eo(n))),{x:a.left+l.scrollLeft-d.x,y:a.top+l.scrollTop-d.y,width:a.width,height:a.height}}function Ln(o){var t=new Map,e=new Set,r=[];o.forEach(function(n){t.set(n.name,n)});function i(n){e.add(n.name);var a=[].concat(n.requires||[],n.requiresIfExists||[]);a.forEach(function(l){if(!e.has(l)){var d=t.get(l);d&&i(d)}}),r.push(n)}return o.forEach(function(n){e.has(n.name)||i(n)}),r}function Cn(o){var t=Ln(o);return Wr.reduce(function(e,r){return e.concat(t.filter(function(i){return i.phase===r}))},[])}function Dn(o){var t;return function(){return t||(t=new Promise(function(e){Promise.resolve().then(function(){t=void 0,e(o())})})),t}}function Nn(o){var t=o.reduce(function(e,r){var i=e[r.name];return e[r.name]=i?Object.assign({},i,r,{options:Object.assign({},i.options,r.options),data:Object.assign({},i.data,r.data)}):r,e},{});return Object.keys(t).map(function(e){return t[e]})}var qo={placement:"bottom",modifiers:[],strategy:"absolute"};function Qo(){for(var o=arguments.length,t=new Array(o),e=0;e<o;e++)t[e]=arguments[e];return!t.some(function(r){return!(r&&typeof r.getBoundingClientRect=="function")})}function Me(o){o===void 0&&(o={});var t=o,e=t.defaultModifiers,r=e===void 0?[]:e,i=t.defaultOptions,n=i===void 0?qo:i;return function(l,d,c){c===void 0&&(c=n);var m={placement:"bottom",orderedModifiers:[],options:Object.assign({},qo,n),modifiersData:{},elements:{reference:l,popper:d},attributes:{},styles:{}},p=[],b=!1,g={state:m,setOptions:function(E){var T=typeof E=="function"?E(m.options):E;u(),m.options=Object.assign({},n,m.options,T),m.scrollParents={reference:Ct(l)?Zt(l):l.contextElement?Zt(l.contextElement):[],popper:Zt(d)};var A=Cn(Nn([].concat(r,m.options.modifiers)));return m.orderedModifiers=A.filter(function(h){return h.enabled}),x(),g.update()},forceUpdate:function(){if(!b){var E=m.elements,T=E.reference,A=E.popper;if(!!Qo(T,A)){m.rects={reference:On(T,ne(A),m.options.strategy==="fixed"),popper:vo(A)},m.reset=!1,m.placement=m.options.placement,m.orderedModifiers.forEach(function(L){return m.modifiersData[L.name]=Object.assign({},L.data)});for(var h=0;h<m.orderedModifiers.length;h++){if(m.reset===!0){m.reset=!1,h=-1;continue}var _=m.orderedModifiers[h],w=_.fn,k=_.options,S=k===void 0?{}:k,D=_.name;typeof w=="function"&&(m=w({state:m,options:S,name:D,instance:g})||m)}}}},update:Dn(function(){return new Promise(function(v){g.forceUpdate(),v(m)})}),destroy:function(){u(),b=!0}};if(!Qo(l,d))return g;g.setOptions(c).then(function(v){!b&&c.onFirstUpdate&&c.onFirstUpdate(v)});function x(){m.orderedModifiers.forEach(function(v){var E=v.name,T=v.options,A=T===void 0?{}:T,h=v.effect;if(typeof h=="function"){var _=h({state:m,name:E,instance:g,options:A}),w=function(){};p.push(_||w)}})}function u(){p.forEach(function(v){return v()}),p=[]}return g}}var $n=Me(),In=[_o,Ao,yo,xo],zn=Me({defaultModifiers:In}),Mn=[_o,Ao,yo,xo,Zr,Qr,ti,Gr,Jr],So=Me({defaultModifiers:Mn}),ei=Object.freeze(Object.defineProperty({__proto__:null,popperGenerator:Me,detectOverflow:It,createPopperBase:$n,createPopper:So,createPopperLite:zn,top:$,bottom:V,right:H,left:I,auto:Ie,basePlacements:Mt,start:yt,end:Lt,clippingParents:Nr,viewport:uo,popper:Tt,reference:$r,variationPlacements:eo,placements:bo,beforeRead:Ir,read:zr,afterRead:Mr,beforeMain:Rr,main:Pr,afterMain:jr,beforeWrite:Vr,write:Hr,afterWrite:Br,modifierPhases:Wr,applyStyles:xo,arrow:Gr,computeStyles:yo,eventListeners:_o,flip:Qr,hide:Jr,offset:Zr,popperOffsets:Ao,preventOverflow:ti},Symbol.toStringTag,{value:"Module"}));/*!
+  * Bootstrap v5.0.0 (https://getbootstrap.com/)
+  * Copyright 2011-2021 The Bootstrap Authors (https://github.com/twbs/bootstrap/graphs/contributors)
+  * Licensed under MIT (https://github.com/twbs/bootstrap/blob/main/LICENSE)
+  */const Rn=1e6,Pn=1e3,ro="transitionend",jn=o=>o==null?`${o}`:{}.toString.call(o).match(/\s([a-z]+)/i)[1].toLowerCase(),oi=o=>{do o+=Math.floor(Math.random()*Rn);while(document.getElementById(o));return o},ri=o=>{let t=o.getAttribute("data-bs-target");if(!t||t==="#"){let e=o.getAttribute("href");if(!e||!e.includes("#")&&!e.startsWith("."))return null;e.includes("#")&&!e.startsWith("#")&&(e=`#${e.split("#")[1]}`),t=e&&e!=="#"?e.trim():null}return t},Oo=o=>{const t=ri(o);return t&&document.querySelector(t)?t:null},lt=o=>{const t=ri(o);return t?document.querySelector(t):null},z=o=>{if(!o)return 0;let{transitionDuration:t,transitionDelay:e}=window.getComputedStyle(o);const r=Number.parseFloat(t),i=Number.parseFloat(e);return!r&&!i?0:(t=t.split(",")[0],e=e.split(",")[0],(Number.parseFloat(t)+Number.parseFloat(e))*Pn)},ii=o=>{o.dispatchEvent(new Event(ro))},_t=o=>(o[0]||o).nodeType,N=(o,t)=>{let e=!1;const i=t+5;function n(){e=!0,o.removeEventListener(ro,n)}o.addEventListener(ro,n),setTimeout(()=>{e||ii(o)},i)},ot=(o,t,e)=>{Object.keys(e).forEach(r=>{const i=e[r],n=t[r],a=n&&_t(n)?"element":jn(n);if(!new RegExp(i).test(a))throw new TypeError(`${o.toUpperCase()}: Option "${r}" provided type "${a}" but expected type "${i}".`)})},Re=o=>{if(!o)return!1;if(o.style&&o.parentNode&&o.parentNode.style){const t=getComputedStyle(o),e=getComputedStyle(o.parentNode);return t.display!=="none"&&e.display!=="none"&&t.visibility!=="hidden"}return!1},At=o=>!o||o.nodeType!==Node.ELEMENT_NODE||o.classList.contains("disabled")?!0:typeof o.disabled!="undefined"?o.disabled:o.hasAttribute("disabled")&&o.getAttribute("disabled")!=="false",ni=o=>{if(!document.documentElement.attachShadow)return null;if(typeof o.getRootNode=="function"){const t=o.getRootNode();return t instanceof ShadowRoot?t:null}return o instanceof ShadowRoot?o:o.parentNode?ni(o.parentNode):null},Le=()=>{},Rt=o=>o.offsetHeight,ai=()=>{const{jQuery:o}=window;return o&&!document.body.hasAttribute("data-bs-no-jquery")?o:null},Vn=o=>{document.readyState==="loading"?document.addEventListener("DOMContentLoaded",o):o()},M=()=>document.documentElement.dir==="rtl",Y=(o,t)=>{Vn(()=>{const e=ai();if(e){const r=e.fn[o];e.fn[o]=t.jQueryInterface,e.fn[o].Constructor=t,e.fn[o].noConflict=()=>(e.fn[o]=r,t.jQueryInterface)}})},gt=o=>{typeof o=="function"&&o()},nt=new Map;var y={set(o,t,e){nt.has(o)||nt.set(o,new Map);const r=nt.get(o);if(!r.has(t)&&r.size!==0){console.error(`Bootstrap doesn't allow more than one instance per element. Bound instance: ${Array.from(r.keys())[0]}.`);return}r.set(t,e)},get(o,t){return nt.has(o)&&nt.get(o).get(t)||null},remove(o,t){if(!nt.has(o))return;const e=nt.get(o);e.delete(t),e.size===0&&nt.delete(o)}};const Hn=/[^.]*(?=\..*)\.|.*/,Bn=/\..*/,Wn=/::\d+$/,Ye={};let Jo=1;const Kn={mouseenter:"mouseover",mouseleave:"mouseout"},Yn=/^(mouseenter|mouseleave)/i,li=new Set(["click","dblclick","mouseup","mousedown","contextmenu","mousewheel","DOMMouseScroll","mouseover","mouseout","mousemove","selectstart","selectend","keydown","keypress","keyup","orientationchange","touchstart","touchmove","touchend","touchcancel","pointerdown","pointermove","pointerup","pointerleave","pointercancel","gesturestart","gesturechange","gestureend","focus","blur","change","reset","select","submit","focusin","focusout","load","unload","beforeunload","resize","move","DOMContentLoaded","readystatechange","error","abort","scroll"]);function si(o,t){return t&&`${t}::${Jo++}`||o.uidEvent||Jo++}function di(o){const t=si(o);return o.uidEvent=t,Ye[t]=Ye[t]||{},Ye[t]}function Un(o,t){return function e(r){return r.delegateTarget=o,e.oneOff&&s.off(o,r.type,t),t.apply(o,[r])}}function Fn(o,t,e){return function r(i){const n=o.querySelectorAll(t);for(let{target:a}=i;a&&a!==this;a=a.parentNode)for(let l=n.length;l--;)if(n[l]===a)return i.delegateTarget=a,r.oneOff&&s.off(o,i.type,t,e),e.apply(a,[i]);return null}}function mi(o,t,e=null){const r=Object.keys(o);for(let i=0,n=r.length;i<n;i++){const a=o[r[i]];if(a.originalHandler===t&&a.delegationSelector===e)return a}return null}function ci(o,t,e){const r=typeof t=="string",i=r?e:t;let n=pi(o);return li.has(n)||(n=o),[r,i,n]}function Zo(o,t,e,r,i){if(typeof t!="string"||!o)return;if(e||(e=r,r=null),Yn.test(t)){const g=x=>function(u){if(!u.relatedTarget||u.relatedTarget!==u.delegateTarget&&!u.delegateTarget.contains(u.relatedTarget))return x.call(this,u)};r?r=g(r):e=g(e)}const[n,a,l]=ci(t,e,r),d=di(o),c=d[l]||(d[l]={}),m=mi(c,a,n?e:null);if(m){m.oneOff=m.oneOff&&i;return}const p=si(a,t.replace(Hn,"")),b=n?Fn(o,e,r):Un(o,e);b.delegationSelector=n?e:null,b.originalHandler=a,b.oneOff=i,b.uidEvent=p,c[p]=b,o.addEventListener(l,b,n)}function io(o,t,e,r,i){const n=mi(t[e],r,i);!n||(o.removeEventListener(e,n,Boolean(i)),delete t[e][n.uidEvent])}function Gn(o,t,e,r){const i=t[e]||{};Object.keys(i).forEach(n=>{if(n.includes(r)){const a=i[n];io(o,t,e,a.originalHandler,a.delegationSelector)}})}function pi(o){return o=o.replace(Bn,""),Kn[o]||o}const s={on(o,t,e,r){Zo(o,t,e,r,!1)},one(o,t,e,r){Zo(o,t,e,r,!0)},off(o,t,e,r){if(typeof t!="string"||!o)return;const[i,n,a]=ci(t,e,r),l=a!==t,d=di(o),c=t.startsWith(".");if(typeof n!="undefined"){if(!d||!d[a])return;io(o,d,a,n,i?e:null);return}c&&Object.keys(d).forEach(p=>{Gn(o,d,p,t.slice(1))});const m=d[a]||{};Object.keys(m).forEach(p=>{const b=p.replace(Wn,"");if(!l||t.includes(b)){const g=m[p];io(o,d,a,g.originalHandler,g.delegationSelector)}})},trigger(o,t,e){if(typeof t!="string"||!o)return null;const r=ai(),i=pi(t),n=t!==i,a=li.has(i);let l,d=!0,c=!0,m=!1,p=null;return n&&r&&(l=r.Event(t,e),r(o).trigger(l),d=!l.isPropagationStopped(),c=!l.isImmediatePropagationStopped(),m=l.isDefaultPrevented()),a?(p=document.createEvent("HTMLEvents"),p.initEvent(i,d,!0)):p=new CustomEvent(t,{bubbles:d,cancelable:!0}),typeof e!="undefined"&&Object.keys(e).forEach(b=>{Object.defineProperty(p,b,{get(){return e[b]}})}),m&&p.preventDefault(),c&&o.dispatchEvent(p),p.defaultPrevented&&typeof l!="undefined"&&l.preventDefault(),p}},Xn="5.0.0";class Q{constructor(t){t=typeof t=="string"?document.querySelector(t):t,t&&(this._element=t,y.set(this._element,this.constructor.DATA_KEY,this))}dispose(){y.remove(this._element,this.constructor.DATA_KEY),s.off(this._element,`.${this.constructor.DATA_KEY}`),this._element=null}static getInstance(t){return y.get(t,this.DATA_KEY)}static get VERSION(){return Xn}}const qn="alert",no="bs.alert",Lo=`.${no}`,Qn=".data-api",Jn='[data-bs-dismiss="alert"]',Zn=`close${Lo}`,ta=`closed${Lo}`,ea=`click${Lo}${Qn}`,oa="alert",ra="fade",ia="show";class oe extends Q{static get DATA_KEY(){return no}close(t){const e=t?this._getRootElement(t):this._element,r=this._triggerCloseEvent(e);r===null||r.defaultPrevented||this._removeElement(e)}_getRootElement(t){return lt(t)||t.closest(`.${oa}`)}_triggerCloseEvent(t){return s.trigger(t,Zn)}_removeElement(t){if(t.classList.remove(ia),!t.classList.contains(ra)){this._destroyElement(t);return}const e=z(t);s.one(t,"transitionend",()=>this._destroyElement(t)),N(t,e)}_destroyElement(t){t.parentNode&&t.parentNode.removeChild(t),s.trigger(t,ta)}static jQueryInterface(t){return this.each(function(){let e=y.get(this,no);e||(e=new oe(this)),t==="close"&&e[t](this)})}static handleDismiss(t){return function(e){e&&e.preventDefault(),t.close(this)}}}s.on(document,ea,Jn,oe.handleDismiss(new oe));Y(qn,oe);const na="button",Ce="bs.button",aa=`.${Ce}`,la=".data-api",sa="active",tr='[data-bs-toggle="button"]',da=`click${aa}${la}`;class Pe extends Q{static get DATA_KEY(){return Ce}toggle(){this._element.setAttribute("aria-pressed",this._element.classList.toggle(sa))}static jQueryInterface(t){return this.each(function(){let e=y.get(this,Ce);e||(e=new Pe(this)),t==="toggle"&&e[t]()})}}s.on(document,da,tr,o=>{o.preventDefault();const t=o.target.closest(tr);let e=y.get(t,Ce);e||(e=new Pe(t)),e.toggle()});Y(na,Pe);function er(o){return o==="true"?!0:o==="false"?!1:o===Number(o).toString()?Number(o):o===""||o==="null"?null:o}function Ue(o){return o.replace(/[A-Z]/g,t=>`-${t.toLowerCase()}`)}const O={setDataAttribute(o,t,e){o.setAttribute(`data-bs-${Ue(t)}`,e)},removeDataAttribute(o,t){o.removeAttribute(`data-bs-${Ue(t)}`)},getDataAttributes(o){if(!o)return{};const t={};return Object.keys(o.dataset).filter(e=>e.startsWith("bs")).forEach(e=>{let r=e.replace(/^bs/,"");r=r.charAt(0).toLowerCase()+r.slice(1,r.length),t[r]=er(o.dataset[e])}),t},getDataAttribute(o,t){return er(o.getAttribute(`data-bs-${Ue(t)}`))},offset(o){const t=o.getBoundingClientRect();return{top:t.top+document.body.scrollTop,left:t.left+document.body.scrollLeft}},position(o){return{top:o.offsetTop,left:o.offsetLeft}}},ma=3,f={find(o,t=document.documentElement){return[].concat(...Element.prototype.querySelectorAll.call(t,o))},findOne(o,t=document.documentElement){return Element.prototype.querySelector.call(t,o)},children(o,t){return[].concat(...o.children).filter(e=>e.matches(t))},parents(o,t){const e=[];let r=o.parentNode;for(;r&&r.nodeType===Node.ELEMENT_NODE&&r.nodeType!==ma;)r.matches(t)&&e.push(r),r=r.parentNode;return e},prev(o,t){let e=o.previousElementSibling;for(;e;){if(e.matches(t))return[e];e=e.previousElementSibling}return[]},next(o,t){let e=o.nextElementSibling;for(;e;){if(e.matches(t))return[e];e=e.nextElementSibling}return[]}},gi="carousel",te="bs.carousel",B=`.${te}`,fi=".data-api",ca="ArrowLeft",pa="ArrowRight",ga=500,fa=40,Fe={interval:5e3,keyboard:!0,slide:!1,pause:"hover",wrap:!0,touch:!0},ua={interval:"(number|boolean)",keyboard:"boolean",slide:"(boolean|string)",pause:"(string|boolean)",wrap:"boolean",touch:"boolean"},ft="next",at="prev",ut="left",Wt="right",ba=`slide${B}`,Ge=`slid${B}`,ha=`keydown${B}`,xa=`mouseenter${B}`,va=`mouseleave${B}`,wa=`touchstart${B}`,ya=`touchmove${B}`,_a=`touchend${B}`,ka=`pointerdown${B}`,Ea=`pointerup${B}`,Ta=`dragstart${B}`,Aa=`load${B}${fi}`,Sa=`click${B}${fi}`,Oa="carousel",bt="active",La="slide",Ca="carousel-item-end",Da="carousel-item-start",Na="carousel-item-next",$a="carousel-item-prev",Ia="pointer-event",za=".active",ue=".active.carousel-item",Ma=".carousel-item",Ra=".carousel-item img",Pa=".carousel-item-next, .carousel-item-prev",ja=".carousel-indicators",Va="[data-bs-target]",Ha="[data-bs-slide], [data-bs-slide-to]",Ba='[data-bs-ride="carousel"]',or="touch",rr="pen";class wt extends Q{constructor(t,e){super(t),this._items=null,this._interval=null,this._activeElement=null,this._isPaused=!1,this._isSliding=!1,this.touchTimeout=null,this.touchStartX=0,this.touchDeltaX=0,this._config=this._getConfig(e),this._indicatorsElement=f.findOne(ja,this._element),this._touchSupported="ontouchstart"in document.documentElement||navigator.maxTouchPoints>0,this._pointerEvent=Boolean(window.PointerEvent),this._addEventListeners()}static get Default(){return Fe}static get DATA_KEY(){return te}next(){this._isSliding||this._slide(ft)}nextWhenVisible(){!document.hidden&&Re(this._element)&&this.next()}prev(){this._isSliding||this._slide(at)}pause(t){t||(this._isPaused=!0),f.findOne(Pa,this._element)&&(ii(this._element),this.cycle(!0)),clearInterval(this._interval),this._interval=null}cycle(t){t||(this._isPaused=!1),this._interval&&(clearInterval(this._interval),this._interval=null),this._config&&this._config.interval&&!this._isPaused&&(this._updateInterval(),this._interval=setInterval((document.visibilityState?this.nextWhenVisible:this.next).bind(this),this._config.interval))}to(t){this._activeElement=f.findOne(ue,this._element);const e=this._getItemIndex(this._activeElement);if(t>this._items.length-1||t<0)return;if(this._isSliding){s.one(this._element,Ge,()=>this.to(t));return}if(e===t){this.pause(),this.cycle();return}const r=t>e?ft:at;this._slide(r,this._items[t])}dispose(){this._items=null,this._config=null,this._interval=null,this._isPaused=null,this._isSliding=null,this._activeElement=null,this._indicatorsElement=null,super.dispose()}_getConfig(t){return t={...Fe,...t},ot(gi,t,ua),t}_handleSwipe(){const t=Math.abs(this.touchDeltaX);if(t<=fa)return;const e=t/this.touchDeltaX;this.touchDeltaX=0,e&&this._slide(e>0?Wt:ut)}_addEventListeners(){this._config.keyboard&&s.on(this._element,ha,t=>this._keydown(t)),this._config.pause==="hover"&&(s.on(this._element,xa,t=>this.pause(t)),s.on(this._element,va,t=>this.cycle(t))),this._config.touch&&this._touchSupported&&this._addTouchEventListeners()}_addTouchEventListeners(){const t=i=>{this._pointerEvent&&(i.pointerType===rr||i.pointerType===or)?this.touchStartX=i.clientX:this._pointerEvent||(this.touchStartX=i.touches[0].clientX)},e=i=>{this.touchDeltaX=i.touches&&i.touches.length>1?0:i.touches[0].clientX-this.touchStartX},r=i=>{this._pointerEvent&&(i.pointerType===rr||i.pointerType===or)&&(this.touchDeltaX=i.clientX-this.touchStartX),this._handleSwipe(),this._config.pause==="hover"&&(this.pause(),this.touchTimeout&&clearTimeout(this.touchTimeout),this.touchTimeout=setTimeout(n=>this.cycle(n),ga+this._config.interval))};f.find(Ra,this._element).forEach(i=>{s.on(i,Ta,n=>n.preventDefault())}),this._pointerEvent?(s.on(this._element,ka,i=>t(i)),s.on(this._element,Ea,i=>r(i)),this._element.classList.add(Ia)):(s.on(this._element,wa,i=>t(i)),s.on(this._element,ya,i=>e(i)),s.on(this._element,_a,i=>r(i)))}_keydown(t){/input|textarea/i.test(t.target.tagName)||(t.key===ca?(t.preventDefault(),this._slide(Wt)):t.key===pa&&(t.preventDefault(),this._slide(ut)))}_getItemIndex(t){return this._items=t&&t.parentNode?f.find(Ma,t.parentNode):[],this._items.indexOf(t)}_getItemByOrder(t,e){const r=t===ft,i=t===at,n=this._getItemIndex(e),a=this._items.length-1;if((i&&n===0||r&&n===a)&&!this._config.wrap)return e;const c=(n+(i?-1:1))%this._items.length;return c===-1?this._items[this._items.length-1]:this._items[c]}_triggerSlideEvent(t,e){const r=this._getItemIndex(t),i=this._getItemIndex(f.findOne(ue,this._element));return s.trigger(this._element,ba,{relatedTarget:t,direction:e,from:i,to:r})}_setActiveIndicatorElement(t){if(this._indicatorsElement){const e=f.findOne(za,this._indicatorsElement);e.classList.remove(bt),e.removeAttribute("aria-current");const r=f.find(Va,this._indicatorsElement);for(let i=0;i<r.length;i++)if(Number.parseInt(r[i].getAttribute("data-bs-slide-to"),10)===this._getItemIndex(t)){r[i].classList.add(bt),r[i].setAttribute("aria-current","true");break}}}_updateInterval(){const t=this._activeElement||f.findOne(ue,this._element);if(!t)return;const e=Number.parseInt(t.getAttribute("data-bs-interval"),10);e?(this._config.defaultInterval=this._config.defaultInterval||this._config.interval,this._config.interval=e):this._config.interval=this._config.defaultInterval||this._config.interval}_slide(t,e){const r=this._directionToOrder(t),i=f.findOne(ue,this._element),n=this._getItemIndex(i),a=e||this._getItemByOrder(r,i),l=this._getItemIndex(a),d=Boolean(this._interval),c=r===ft,m=c?Da:Ca,p=c?Na:$a,b=this._orderToDirection(r);if(a&&a.classList.contains(bt)){this._isSliding=!1;return}if(!this._triggerSlideEvent(a,b).defaultPrevented&&!(!i||!a)){if(this._isSliding=!0,d&&this.pause(),this._setActiveIndicatorElement(a),this._activeElement=a,this._element.classList.contains(La)){a.classList.add(p),Rt(a),i.classList.add(m),a.classList.add(m);const x=z(i);s.one(i,"transitionend",()=>{a.classList.remove(m,p),a.classList.add(bt),i.classList.remove(bt,p,m),this._isSliding=!1,setTimeout(()=>{s.trigger(this._element,Ge,{relatedTarget:a,direction:b,from:n,to:l})},0)}),N(i,x)}else i.classList.remove(bt),a.classList.add(bt),this._isSliding=!1,s.trigger(this._element,Ge,{relatedTarget:a,direction:b,from:n,to:l});d&&this.cycle()}}_directionToOrder(t){return[Wt,ut].includes(t)?M()?t===ut?at:ft:t===ut?ft:at:t}_orderToDirection(t){return[ft,at].includes(t)?M()?t===at?ut:Wt:t===at?Wt:ut:t}static carouselInterface(t,e){let r=y.get(t,te),i={...Fe,...O.getDataAttributes(t)};typeof e=="object"&&(i={...i,...e});const n=typeof e=="string"?e:i.slide;if(r||(r=new wt(t,i)),typeof e=="number")r.to(e);else if(typeof n=="string"){if(typeof r[n]=="undefined")throw new TypeError(`No method named "${n}"`);r[n]()}else i.interval&&i.ride&&(r.pause(),r.cycle())}static jQueryInterface(t){return this.each(function(){wt.carouselInterface(this,t)})}static dataApiClickHandler(t){const e=lt(this);if(!e||!e.classList.contains(Oa))return;const r={...O.getDataAttributes(e),...O.getDataAttributes(this)},i=this.getAttribute("data-bs-slide-to");i&&(r.interval=!1),wt.carouselInterface(e,r),i&&y.get(e,te).to(i),t.preventDefault()}}s.on(document,Sa,Ha,wt.dataApiClickHandler);s.on(window,Aa,()=>{const o=f.find(Ba);for(let t=0,e=o.length;t<e;t++)wt.carouselInterface(o[t],y.get(o[t],te))});Y(gi,wt);const ui="collapse",St="bs.collapse",ae=`.${St}`,Wa=".data-api",Xe={toggle:!0,parent:""},Ka={toggle:"boolean",parent:"(string|element)"},Ya=`show${ae}`,Ua=`shown${ae}`,Fa=`hide${ae}`,Ga=`hidden${ae}`,Xa=`click${ae}${Wa}`,ht="show",Kt="collapse",be="collapsing",he="collapsed",ir="width",qa="height",Qa=".show, .collapsing",qt='[data-bs-toggle="collapse"]';class Ot extends Q{constructor(t,e){super(t),this._isTransitioning=!1,this._config=this._getConfig(e),this._triggerArray=f.find(`${qt}[href="#${this._element.id}"],${qt}[data-bs-target="#${this._element.id}"]`);const r=f.find(qt);for(let i=0,n=r.length;i<n;i++){const a=r[i],l=Oo(a),d=f.find(l).filter(c=>c===this._element);l!==null&&d.length&&(this._selector=l,this._triggerArray.push(a))}this._parent=this._config.parent?this._getParent():null,this._config.parent||this._addAriaAndCollapsedClass(this._element,this._triggerArray),this._config.toggle&&this.toggle()}static get Default(){return Xe}static get DATA_KEY(){return St}toggle(){this._element.classList.contains(ht)?this.hide():this.show()}show(){if(this._isTransitioning||this._element.classList.contains(ht))return;let t,e;this._parent&&(t=f.find(Qa,this._parent).filter(m=>typeof this._config.parent=="string"?m.getAttribute("data-bs-parent")===this._config.parent:m.classList.contains(Kt)),t.length===0&&(t=null));const r=f.findOne(this._selector);if(t){const m=t.find(p=>r!==p);if(e=m?y.get(m,St):null,e&&e._isTransitioning)return}if(s.trigger(this._element,Ya).defaultPrevented)return;t&&t.forEach(m=>{r!==m&&Ot.collapseInterface(m,"hide"),e||y.set(m,St,null)});const n=this._getDimension();this._element.classList.remove(Kt),this._element.classList.add(be),this._element.style[n]=0,this._triggerArray.length&&this._triggerArray.forEach(m=>{m.classList.remove(he),m.setAttribute("aria-expanded",!0)}),this.setTransitioning(!0);const a=()=>{this._element.classList.remove(be),this._element.classList.add(Kt,ht),this._element.style[n]="",this.setTransitioning(!1),s.trigger(this._element,Ua)},d=`scroll${n[0].toUpperCase()+n.slice(1)}`,c=z(this._element);s.one(this._element,"transitionend",a),N(this._element,c),this._element.style[n]=`${this._element[d]}px`}hide(){if(this._isTransitioning||!this._element.classList.contains(ht)||s.trigger(this._element,Fa).defaultPrevented)return;const e=this._getDimension();this._element.style[e]=`${this._element.getBoundingClientRect()[e]}px`,Rt(this._element),this._element.classList.add(be),this._element.classList.remove(Kt,ht);const r=this._triggerArray.length;if(r>0)for(let a=0;a<r;a++){const l=this._triggerArray[a],d=lt(l);d&&!d.classList.contains(ht)&&(l.classList.add(he),l.setAttribute("aria-expanded",!1))}this.setTransitioning(!0);const i=()=>{this.setTransitioning(!1),this._element.classList.remove(be),this._element.classList.add(Kt),s.trigger(this._element,Ga)};this._element.style[e]="";const n=z(this._element);s.one(this._element,"transitionend",i),N(this._element,n)}setTransitioning(t){this._isTransitioning=t}dispose(){super.dispose(),this._config=null,this._parent=null,this._triggerArray=null,this._isTransitioning=null}_getConfig(t){return t={...Xe,...t},t.toggle=Boolean(t.toggle),ot(ui,t,Ka),t}_getDimension(){return this._element.classList.contains(ir)?ir:qa}_getParent(){let{parent:t}=this._config;_t(t)?(typeof t.jquery!="undefined"||typeof t[0]!="undefined")&&(t=t[0]):t=f.findOne(t);const e=`${qt}[data-bs-parent="${t}"]`;return f.find(e,t).forEach(r=>{const i=lt(r);this._addAriaAndCollapsedClass(i,[r])}),t}_addAriaAndCollapsedClass(t,e){if(!t||!e.length)return;const r=t.classList.contains(ht);e.forEach(i=>{r?i.classList.remove(he):i.classList.add(he),i.setAttribute("aria-expanded",r)})}static collapseInterface(t,e){let r=y.get(t,St);const i={...Xe,...O.getDataAttributes(t),...typeof e=="object"&&e?e:{}};if(!r&&i.toggle&&typeof e=="string"&&/show|hide/.test(e)&&(i.toggle=!1),r||(r=new Ot(t,i)),typeof e=="string"){if(typeof r[e]=="undefined")throw new TypeError(`No method named "${e}"`);r[e]()}}static jQueryInterface(t){return this.each(function(){Ot.collapseInterface(this,t)})}}s.on(document,Xa,qt,function(o){(o.target.tagName==="A"||o.delegateTarget&&o.delegateTarget.tagName==="A")&&o.preventDefault();const t=O.getDataAttributes(this),e=Oo(this);f.find(e).forEach(i=>{const n=y.get(i,St);let a;n?(n._parent===null&&typeof t.parent=="string"&&(n._config.parent=t.parent,n._parent=n._getParent()),a="toggle"):a=t,Ot.collapseInterface(i,a)})});Y(ui,Ot);const ao="dropdown",ke="bs.dropdown",dt=`.${ke}`,Co=".data-api",Ee="Escape",nr="Space",ar="Tab",Te="ArrowUp",Ae="ArrowDown",Ja=2,Za=new RegExp(`${Te}|${Ae}|${Ee}`),tl=`hide${dt}`,el=`hidden${dt}`,ol=`show${dt}`,rl=`shown${dt}`,il=`click${dt}`,bi=`click${dt}${Co}`,hi=`keydown${dt}${Co}`,nl=`keyup${dt}${Co}`,tt="show",al="dropup",ll="dropend",sl="dropstart",dl="navbar",ee='[data-bs-toggle="dropdown"]',lo=".dropdown-menu",ml=".navbar-nav",cl=".dropdown-menu .dropdown-item:not(.disabled):not(:disabled)",pl=M()?"top-end":"top-start",gl=M()?"top-start":"top-end",fl=M()?"bottom-end":"bottom-start",ul=M()?"bottom-start":"bottom-end",bl=M()?"left-start":"right-start",hl=M()?"right-start":"left-start",xl={offset:[0,2],boundary:"clippingParents",reference:"toggle",display:"dynamic",popperConfig:null,autoClose:!0},vl={offset:"(array|string|function)",boundary:"(string|element)",reference:"(string|element|object)",display:"string",popperConfig:"(null|object|function)",autoClose:"(boolean|string)"};class j extends Q{constructor(t,e){super(t),this._popper=null,this._config=this._getConfig(e),this._menu=this._getMenuElement(),this._inNavbar=this._detectNavbar(),this._addEventListeners()}static get Default(){return xl}static get DefaultType(){return vl}static get DATA_KEY(){return ke}toggle(){if(At(this._element))return;if(this._element.classList.contains(tt)){this.hide();return}this.show()}show(){if(At(this._element)||this._menu.classList.contains(tt))return;const t=j.getParentFromElement(this._element),e={relatedTarget:this._element};if(!s.trigger(this._element,ol,e).defaultPrevented){if(this._inNavbar)O.setDataAttribute(this._menu,"popper","none");else{if(typeof ei=="undefined")throw new TypeError("Bootstrap's dropdowns require Popper (https://popper.js.org)");let i=this._element;this._config.reference==="parent"?i=t:_t(this._config.reference)?(i=this._config.reference,typeof this._config.reference.jquery!="undefined"&&(i=this._config.reference[0])):typeof this._config.reference=="object"&&(i=this._config.reference);const n=this._getPopperConfig(),a=n.modifiers.find(l=>l.name==="applyStyles"&&l.enabled===!1);this._popper=So(i,this._menu,n),a&&O.setDataAttribute(this._menu,"popper","static")}"ontouchstart"in document.documentElement&&!t.closest(ml)&&[].concat(...document.body.children).forEach(i=>s.on(i,"mouseover",Le)),this._element.focus(),this._element.setAttribute("aria-expanded",!0),this._menu.classList.toggle(tt),this._element.classList.toggle(tt),s.trigger(this._element,rl,e)}}hide(){if(At(this._element)||!this._menu.classList.contains(tt))return;const t={relatedTarget:this._element};this._completeHide(t)}dispose(){this._menu=null,this._popper&&(this._popper.destroy(),this._popper=null),super.dispose()}update(){this._inNavbar=this._detectNavbar(),this._popper&&this._popper.update()}_addEventListeners(){s.on(this._element,il,t=>{t.preventDefault(),this.toggle()})}_completeHide(t){s.trigger(this._element,tl,t).defaultPrevented||("ontouchstart"in document.documentElement&&[].concat(...document.body.children).forEach(r=>s.off(r,"mouseover",Le)),this._popper&&this._popper.destroy(),this._menu.classList.remove(tt),this._element.classList.remove(tt),this._element.setAttribute("aria-expanded","false"),O.removeDataAttribute(this._menu,"popper"),s.trigger(this._element,el,t))}_getConfig(t){if(t={...this.constructor.Default,...O.getDataAttributes(this._element),...t},ot(ao,t,this.constructor.DefaultType),typeof t.reference=="object"&&!_t(t.reference)&&typeof t.reference.getBoundingClientRect!="function")throw new TypeError(`${ao.toUpperCase()}: Option "reference" provided type "object" without a required "getBoundingClientRect" method.`);return t}_getMenuElement(){return f.next(this._element,lo)[0]}_getPlacement(){const t=this._element.parentNode;if(t.classList.contains(ll))return bl;if(t.classList.contains(sl))return hl;const e=getComputedStyle(this._menu).getPropertyValue("--bs-position").trim()==="end";return t.classList.contains(al)?e?gl:pl:e?ul:fl}_detectNavbar(){return this._element.closest(`.${dl}`)!==null}_getOffset(){const{offset:t}=this._config;return typeof t=="string"?t.split(",").map(e=>Number.parseInt(e,10)):typeof t=="function"?e=>t(e,this._element):t}_getPopperConfig(){const t={placement:this._getPlacement(),modifiers:[{name:"preventOverflow",options:{boundary:this._config.boundary}},{name:"offset",options:{offset:this._getOffset()}}]};return this._config.display==="static"&&(t.modifiers=[{name:"applyStyles",enabled:!1}]),{...t,...typeof this._config.popperConfig=="function"?this._config.popperConfig(t):this._config.popperConfig}}_selectMenuItem(t){const e=f.find(cl,this._menu).filter(Re);if(!e.length)return;let r=e.indexOf(t.target);t.key===Te&&r>0&&r--,t.key===Ae&&r<e.length-1&&r++,r=r===-1?0:r,e[r].focus()}static dropdownInterface(t,e){let r=y.get(t,ke);const i=typeof e=="object"?e:null;if(r||(r=new j(t,i)),typeof e=="string"){if(typeof r[e]=="undefined")throw new TypeError(`No method named "${e}"`);r[e]()}}static jQueryInterface(t){return this.each(function(){j.dropdownInterface(this,t)})}static clearMenus(t){if(t&&(t.button===Ja||t.type==="keyup"&&t.key!==ar||/input|select|option|textarea|form/i.test(t.target.tagName)))return;const e=f.find(ee);for(let r=0,i=e.length;r<i;r++){const n=y.get(e[r],ke);if(!n||n._config.autoClose===!1||!n._element.classList.contains(tt))continue;const a={relatedTarget:n._element};if(t){const l=t.composedPath(),d=l.includes(n._menu);if(l.includes(n._element)||n._config.autoClose==="inside"&&!d||n._config.autoClose==="outside"&&d||t.type==="keyup"&&t.key===ar&&n._menu.contains(t.target))continue;t.type==="click"&&(a.clickEvent=t)}n._completeHide(a)}}static getParentFromElement(t){return lt(t)||t.parentNode}static dataApiKeydownHandler(t){if(/input|textarea/i.test(t.target.tagName)?t.key===nr||t.key!==Ee&&(t.key!==Ae&&t.key!==Te||t.target.closest(lo)):!Za.test(t.key))return;const e=this.classList.contains(tt);if(!e&&t.key===Ee||(t.preventDefault(),t.stopPropagation(),At(this)))return;const r=()=>this.matches(ee)?this:f.prev(this,ee)[0];if(t.key===Ee){r().focus(),j.clearMenus();return}if(!e&&(t.key===Te||t.key===Ae)){r().click();return}if(!e||t.key===nr){j.clearMenus();return}j.getInstance(r())._selectMenuItem(t)}}s.on(document,hi,ee,j.dataApiKeydownHandler);s.on(document,hi,lo,j.dataApiKeydownHandler);s.on(document,bi,j.clearMenus);s.on(document,nl,j.clearMenus);s.on(document,bi,ee,function(o){o.preventDefault(),j.dropdownInterface(this)});Y(ao,j);const xi=".fixed-top, .fixed-bottom, .is-fixed, .sticky-top",vi=".sticky-top",Do=()=>{const o=document.documentElement.clientWidth;return Math.abs(window.innerWidth-o)},wi=(o=Do())=>{wl(),qe("body","paddingRight",t=>t+o),qe(xi,"paddingRight",t=>t+o),qe(vi,"marginRight",t=>t-o)},wl=()=>{const o=document.body.style.overflow;o&&O.setDataAttribute(document.body,"overflow",o),document.body.style.overflow="hidden"},qe=(o,t,e)=>{const r=Do();f.find(o).forEach(i=>{if(i!==document.body&&window.innerWidth>i.clientWidth+r)return;const n=i.style[t],a=window.getComputedStyle(i)[t];O.setDataAttribute(i,t,n),i.style[t]=`${e(Number.parseFloat(a))}px`})},yi=()=>{xe("body","overflow"),xe("body","paddingRight"),xe(xi,"paddingRight"),xe(vi,"marginRight")},xe=(o,t)=>{f.find(o).forEach(e=>{const r=O.getDataAttribute(e,t);typeof r=="undefined"?e.style.removeProperty(t):(O.removeDataAttribute(e,t),e.style[t]=r)})},yl={isVisible:!0,isAnimated:!1,rootElement:document.body,clickCallback:null},_l={isVisible:"boolean",isAnimated:"boolean",rootElement:"element",clickCallback:"(function|null)"},_i="backdrop",kl="modal-backdrop",El="fade",lr="show",sr=`mousedown.bs.${_i}`;class ki{constructor(t){this._config=this._getConfig(t),this._isAppended=!1,this._element=null}show(t){if(!this._config.isVisible){gt(t);return}this._append(),this._config.isAnimated&&Rt(this._getElement()),this._getElement().classList.add(lr),this._emulateAnimation(()=>{gt(t)})}hide(t){if(!this._config.isVisible){gt(t);return}this._getElement().classList.remove(lr),this._emulateAnimation(()=>{this.dispose(),gt(t)})}_getElement(){if(!this._element){const t=document.createElement("div");t.className=kl,this._config.isAnimated&&t.classList.add(El),this._element=t}return this._element}_getConfig(t){return t={...yl,...typeof t=="object"?t:{}},ot(_i,t,_l),t}_append(){this._isAppended||(this._config.rootElement.appendChild(this._getElement()),s.on(this._getElement(),sr,()=>{gt(this._config.clickCallback)}),this._isAppended=!0)}dispose(){!this._isAppended||(s.off(this._element,sr),this._getElement().parentNode.removeChild(this._element),this._isAppended=!1)}_emulateAnimation(t){if(!this._config.isAnimated){gt(t);return}const e=z(this._getElement());s.one(this._getElement(),"transitionend",()=>gt(t)),N(this._getElement(),e)}}const Ei="modal",Ti="bs.modal",W=`.${Ti}`,Tl=".data-api",dr="Escape",mr={backdrop:!0,keyboard:!0,focus:!0},Al={backdrop:"(boolean|string)",keyboard:"boolean",focus:"boolean"},Sl=`hide${W}`,Ol=`hidePrevented${W}`,Ai=`hidden${W}`,Si=`show${W}`,Ll=`shown${W}`,ve=`focusin${W}`,cr=`resize${W}`,Qe=`click.dismiss${W}`,pr=`keydown.dismiss${W}`,Cl=`mouseup.dismiss${W}`,gr=`mousedown.dismiss${W}`,Dl=`click${W}${Tl}`,fr="modal-open",Nl="fade",ur="show",br="modal-static",$l=".modal-dialog",Il=".modal-body",zl='[data-bs-toggle="modal"]',Ml='[data-bs-dismiss="modal"]';class kt extends Q{constructor(t,e){super(t),this._config=this._getConfig(e),this._dialog=f.findOne($l,this._element),this._backdrop=this._initializeBackDrop(),this._isShown=!1,this._ignoreBackdropClick=!1,this._isTransitioning=!1}static get Default(){return mr}static get DATA_KEY(){return Ti}toggle(t){return this._isShown?this.hide():this.show(t)}show(t){if(this._isShown||this._isTransitioning)return;this._isAnimated()&&(this._isTransitioning=!0);const e=s.trigger(this._element,Si,{relatedTarget:t});this._isShown||e.defaultPrevented||(this._isShown=!0,wi(),document.body.classList.add(fr),this._adjustDialog(),this._setEscapeEvent(),this._setResizeEvent(),s.on(this._element,Qe,Ml,r=>this.hide(r)),s.on(this._dialog,gr,()=>{s.one(this._element,Cl,r=>{r.target===this._element&&(this._ignoreBackdropClick=!0)})}),this._showBackdrop(()=>this._showElement(t)))}hide(t){if(t&&t.preventDefault(),!this._isShown||this._isTransitioning||s.trigger(this._element,Sl).defaultPrevented)return;this._isShown=!1;const r=this._isAnimated();if(r&&(this._isTransitioning=!0),this._setEscapeEvent(),this._setResizeEvent(),s.off(document,ve),this._element.classList.remove(ur),s.off(this._element,Qe),s.off(this._dialog,gr),r){const i=z(this._element);s.one(this._element,"transitionend",n=>this._hideModal(n)),N(this._element,i)}else this._hideModal()}dispose(){[window,this._dialog].forEach(t=>s.off(t,W)),super.dispose(),s.off(document,ve),this._config=null,this._dialog=null,this._backdrop.dispose(),this._backdrop=null,this._isShown=null,this._ignoreBackdropClick=null,this._isTransitioning=null}handleUpdate(){this._adjustDialog()}_initializeBackDrop(){return new ki({isVisible:Boolean(this._config.backdrop),isAnimated:this._isAnimated()})}_getConfig(t){return t={...mr,...O.getDataAttributes(this._element),...t},ot(Ei,t,Al),t}_showElement(t){const e=this._isAnimated(),r=f.findOne(Il,this._dialog);(!this._element.parentNode||this._element.parentNode.nodeType!==Node.ELEMENT_NODE)&&document.body.appendChild(this._element),this._element.style.display="block",this._element.removeAttribute("aria-hidden"),this._element.setAttribute("aria-modal",!0),this._element.setAttribute("role","dialog"),this._element.scrollTop=0,r&&(r.scrollTop=0),e&&Rt(this._element),this._element.classList.add(ur),this._config.focus&&this._enforceFocus();const i=()=>{this._config.focus&&this._element.focus(),this._isTransitioning=!1,s.trigger(this._element,Ll,{relatedTarget:t})};if(e){const n=z(this._dialog);s.one(this._dialog,"transitionend",i),N(this._dialog,n)}else i()}_enforceFocus(){s.off(document,ve),s.on(document,ve,t=>{document!==t.target&&this._element!==t.target&&!this._element.contains(t.target)&&this._element.focus()})}_setEscapeEvent(){this._isShown?s.on(this._element,pr,t=>{this._config.keyboard&&t.key===dr?(t.preventDefault(),this.hide()):!this._config.keyboard&&t.key===dr&&this._triggerBackdropTransition()}):s.off(this._element,pr)}_setResizeEvent(){this._isShown?s.on(window,cr,()=>this._adjustDialog()):s.off(window,cr)}_hideModal(){this._element.style.display="none",this._element.setAttribute("aria-hidden",!0),this._element.removeAttribute("aria-modal"),this._element.removeAttribute("role"),this._isTransitioning=!1,this._backdrop.hide(()=>{document.body.classList.remove(fr),this._resetAdjustments(),yi(),s.trigger(this._element,Ai)})}_showBackdrop(t){s.on(this._element,Qe,e=>{if(this._ignoreBackdropClick){this._ignoreBackdropClick=!1;return}e.target===e.currentTarget&&(this._config.backdrop===!0?this.hide():this._config.backdrop==="static"&&this._triggerBackdropTransition())}),this._backdrop.show(t)}_isAnimated(){return this._element.classList.contains(Nl)}_triggerBackdropTransition(){if(s.trigger(this._element,Ol).defaultPrevented)return;const e=this._element.scrollHeight>document.documentElement.clientHeight;e||(this._element.style.overflowY="hidden"),this._element.classList.add(br);const r=z(this._dialog);s.off(this._element,"transitionend"),s.one(this._element,"transitionend",()=>{this._element.classList.remove(br),e||(s.one(this._element,"transitionend",()=>{this._element.style.overflowY=""}),N(this._element,r))}),N(this._element,r),this._element.focus()}_adjustDialog(){const t=this._element.scrollHeight>document.documentElement.clientHeight,e=Do(),r=e>0;(!r&&t&&!M()||r&&!t&&M())&&(this._element.style.paddingLeft=`${e}px`),(r&&!t&&!M()||!r&&t&&M())&&(this._element.style.paddingRight=`${e}px`)}_resetAdjustments(){this._element.style.paddingLeft="",this._element.style.paddingRight=""}static jQueryInterface(t,e){return this.each(function(){const r=kt.getInstance(this)||new kt(this,typeof t=="object"?t:{});if(typeof t=="string"){if(typeof r[t]=="undefined")throw new TypeError(`No method named "${t}"`);r[t](e)}})}}s.on(document,Dl,zl,function(o){const t=lt(this);["A","AREA"].includes(this.tagName)&&o.preventDefault(),s.one(t,Si,r=>{r.defaultPrevented||s.one(t,Ai,()=>{Re(this)&&this.focus()})}),(kt.getInstance(t)||new kt(t)).toggle(this)});Y(Ei,kt);const Oi="offcanvas",re="bs.offcanvas",rt=`.${re}`,Li=".data-api",Rl=`load${rt}${Li}`,Pl="Escape",hr={backdrop:!0,keyboard:!0,scroll:!1},jl={backdrop:"boolean",keyboard:"boolean",scroll:"boolean"},xr="show",Ci=".offcanvas.show",Vl=`show${rt}`,Hl=`shown${rt}`,Bl=`hide${rt}`,Di=`hidden${rt}`,we=`focusin${rt}`,Wl=`click${rt}${Li}`,Kl=`click.dismiss${rt}`,Yl=`keydown.dismiss${rt}`,Ul='[data-bs-dismiss="offcanvas"]',Fl='[data-bs-toggle="offcanvas"]';class zt extends Q{constructor(t,e){super(t),this._config=this._getConfig(e),this._isShown=!1,this._backdrop=this._initializeBackDrop(),this._addEventListeners()}static get Default(){return hr}static get DATA_KEY(){return re}toggle(t){return this._isShown?this.hide():this.show(t)}show(t){if(this._isShown||s.trigger(this._element,Vl,{relatedTarget:t}).defaultPrevented)return;this._isShown=!0,this._element.style.visibility="visible",this._backdrop.show(),this._config.scroll||(wi(),this._enforceFocusOnElement(this._element)),this._element.removeAttribute("aria-hidden"),this._element.setAttribute("aria-modal",!0),this._element.setAttribute("role","dialog"),this._element.classList.add(xr);const r=()=>{s.trigger(this._element,Hl,{relatedTarget:t})},i=z(this._element);s.one(this._element,"transitionend",r),N(this._element,i)}hide(){if(!this._isShown||s.trigger(this._element,Bl).defaultPrevented)return;s.off(document,we),this._element.blur(),this._isShown=!1,this._element.classList.remove(xr),this._backdrop.hide();const e=()=>{this._element.setAttribute("aria-hidden",!0),this._element.removeAttribute("aria-modal"),this._element.removeAttribute("role"),this._element.style.visibility="hidden",this._config.scroll||yi(),s.trigger(this._element,Di)},r=z(this._element);s.one(this._element,"transitionend",e),N(this._element,r)}dispose(){this._backdrop.dispose(),super.dispose(),s.off(document,we),this._config=null,this._backdrop=null}_getConfig(t){return t={...hr,...O.getDataAttributes(this._element),...typeof t=="object"?t:{}},ot(Oi,t,jl),t}_initializeBackDrop(){return new ki({isVisible:this._config.backdrop,isAnimated:!0,rootElement:this._element.parentNode,clickCallback:()=>this.hide()})}_enforceFocusOnElement(t){s.off(document,we),s.on(document,we,e=>{document!==e.target&&t!==e.target&&!t.contains(e.target)&&t.focus()}),t.focus()}_addEventListeners(){s.on(this._element,Kl,Ul,()=>this.hide()),s.on(this._element,Yl,t=>{this._config.keyboard&&t.key===Pl&&this.hide()})}static jQueryInterface(t){return this.each(function(){const e=y.get(this,re)||new zt(this,typeof t=="object"?t:{});if(typeof t=="string"){if(e[t]===void 0||t.startsWith("_")||t==="constructor")throw new TypeError(`No method named "${t}"`);e[t](this)}})}}s.on(document,Wl,Fl,function(o){const t=lt(this);if(["A","AREA"].includes(this.tagName)&&o.preventDefault(),At(this))return;s.one(t,Di,()=>{Re(this)&&this.focus()});const e=f.findOne(Ci);e&&e!==t&&zt.getInstance(e).hide(),(y.get(t,re)||new zt(t)).toggle(this)});s.on(window,Rl,()=>{f.find(Ci).forEach(o=>(y.get(o,re)||new zt(o)).show())});Y(Oi,zt);const Gl=new Set(["background","cite","href","itemtype","longdesc","poster","src","xlink:href"]),Xl=/^aria-[\w-]*$/i,ql=/^(?:(?:https?|mailto|ftp|tel|file):|[^#&/:?]*(?:[#/?]|$))/i,Ql=/^data:(?:image\/(?:bmp|gif|jpeg|jpg|png|tiff|webp)|video\/(?:mpeg|mp4|ogg|webm)|audio\/(?:mp3|oga|ogg|opus));base64,[\d+/a-z]+=*$/i,Jl=(o,t)=>{const e=o.nodeName.toLowerCase();if(t.includes(e))return Gl.has(e)?Boolean(ql.test(o.nodeValue)||Ql.test(o.nodeValue)):!0;const r=t.filter(i=>i instanceof RegExp);for(let i=0,n=r.length;i<n;i++)if(r[i].test(e))return!0;return!1},Zl={"*":["class","dir","id","lang","role",Xl],a:["target","href","title","rel"],area:[],b:[],br:[],col:[],code:[],div:[],em:[],hr:[],h1:[],h2:[],h3:[],h4:[],h5:[],h6:[],i:[],img:["src","srcset","alt","title","width","height"],li:[],ol:[],p:[],pre:[],s:[],small:[],span:[],sub:[],sup:[],strong:[],u:[],ul:[]};function vr(o,t,e){if(!o.length)return o;if(e&&typeof e=="function")return e(o);const i=new window.DOMParser().parseFromString(o,"text/html"),n=Object.keys(t),a=[].concat(...i.body.querySelectorAll("*"));for(let l=0,d=a.length;l<d;l++){const c=a[l],m=c.nodeName.toLowerCase();if(!n.includes(m)){c.parentNode.removeChild(c);continue}const p=[].concat(...c.attributes),b=[].concat(t["*"]||[],t[m]||[]);p.forEach(g=>{Jl(g,b)||c.removeAttribute(g.nodeName)})}return i.body.innerHTML}const so="tooltip",mo="bs.tooltip",G=`.${mo}`,Ni="bs-tooltip",ts=new RegExp(`(^|\\s)${Ni}\\S+`,"g"),es=new Set(["sanitize","allowList","sanitizeFn"]),os={animation:"boolean",template:"string",title:"(string|element|function)",trigger:"string",delay:"(number|object)",html:"boolean",selector:"(string|boolean)",placement:"(string|function)",offset:"(array|string|function)",container:"(string|element|boolean)",fallbackPlacements:"array",boundary:"(string|element)",customClass:"(string|function)",sanitize:"boolean",sanitizeFn:"(null|function)",allowList:"object",popperConfig:"(null|object|function)"},rs={AUTO:"auto",TOP:"top",RIGHT:M()?"left":"right",BOTTOM:"bottom",LEFT:M()?"right":"left"},is={animation:!0,template:'<div class="tooltip" role="tooltip"><div class="tooltip-arrow"></div><div class="tooltip-inner"></div></div>',trigger:"hover focus",title:"",delay:0,html:!1,selector:!1,placement:"top",offset:[0,0],container:!1,fallbackPlacements:["top","right","bottom","left"],boundary:"clippingParents",customClass:"",sanitize:!0,sanitizeFn:null,allowList:Zl,popperConfig:null},ns={HIDE:`hide${G}`,HIDDEN:`hidden${G}`,SHOW:`show${G}`,SHOWN:`shown${G}`,INSERTED:`inserted${G}`,CLICK:`click${G}`,FOCUSIN:`focusin${G}`,FOCUSOUT:`focusout${G}`,MOUSEENTER:`mouseenter${G}`,MOUSELEAVE:`mouseleave${G}`},ye="fade",wr="modal",Yt="show",Ut="show",Je="out",as=".tooltip-inner",Ft="hover",Ze="focus",ls="click",ss="manual";class Pt extends Q{constructor(t,e){if(typeof ei=="undefined")throw new TypeError("Bootstrap's tooltips require Popper (https://popper.js.org)");super(t),this._isEnabled=!0,this._timeout=0,this._hoverState="",this._activeTrigger={},this._popper=null,this.config=this._getConfig(e),this.tip=null,this._setListeners()}static get Default(){return is}static get NAME(){return so}static get DATA_KEY(){return mo}static get Event(){return ns}static get EVENT_KEY(){return G}static get DefaultType(){return os}enable(){this._isEnabled=!0}disable(){this._isEnabled=!1}toggleEnabled(){this._isEnabled=!this._isEnabled}toggle(t){if(!!this._isEnabled)if(t){const e=this._initializeOnDelegatedTarget(t);e._activeTrigger.click=!e._activeTrigger.click,e._isWithActiveTrigger()?e._enter(null,e):e._leave(null,e)}else{if(this.getTipElement().classList.contains(Yt)){this._leave(null,this);return}this._enter(null,this)}}dispose(){clearTimeout(this._timeout),s.off(this._element.closest(`.${wr}`),"hide.bs.modal",this._hideModalHandler),this.tip&&this.tip.parentNode&&this.tip.parentNode.removeChild(this.tip),this._isEnabled=null,this._timeout=null,this._hoverState=null,this._activeTrigger=null,this._popper&&this._popper.destroy(),this._popper=null,this.config=null,this.tip=null,super.dispose()}show(){if(this._element.style.display==="none")throw new Error("Please use show on visible elements");if(!(this.isWithContent()&&this._isEnabled))return;const t=s.trigger(this._element,this.constructor.Event.SHOW),e=ni(this._element),r=e===null?this._element.ownerDocument.documentElement.contains(this._element):e.contains(this._element);if(t.defaultPrevented||!r)return;const i=this.getTipElement(),n=oi(this.constructor.NAME);i.setAttribute("id",n),this._element.setAttribute("aria-describedby",n),this.setContent(),this.config.animation&&i.classList.add(ye);const a=typeof this.config.placement=="function"?this.config.placement.call(this,i,this._element):this.config.placement,l=this._getAttachment(a);this._addAttachmentClass(l);const d=this._getContainer();y.set(i,this.constructor.DATA_KEY,this),this._element.ownerDocument.documentElement.contains(this.tip)||(d.appendChild(i),s.trigger(this._element,this.constructor.Event.INSERTED)),this._popper?this._popper.update():this._popper=So(this._element,i,this._getPopperConfig(l)),i.classList.add(Yt);const c=typeof this.config.customClass=="function"?this.config.customClass():this.config.customClass;c&&i.classList.add(...c.split(" ")),"ontouchstart"in document.documentElement&&[].concat(...document.body.children).forEach(p=>{s.on(p,"mouseover",Le)});const m=()=>{const p=this._hoverState;this._hoverState=null,s.trigger(this._element,this.constructor.Event.SHOWN),p===Je&&this._leave(null,this)};if(this.tip.classList.contains(ye)){const p=z(this.tip);s.one(this.tip,"transitionend",m),N(this.tip,p)}else m()}hide(){if(!this._popper)return;const t=this.getTipElement(),e=()=>{this._isWithActiveTrigger()||(this._hoverState!==Ut&&t.parentNode&&t.parentNode.removeChild(t),this._cleanTipClass(),this._element.removeAttribute("aria-describedby"),s.trigger(this._element,this.constructor.Event.HIDDEN),this._popper&&(this._popper.destroy(),this._popper=null))};if(!s.trigger(this._element,this.constructor.Event.HIDE).defaultPrevented){if(t.classList.remove(Yt),"ontouchstart"in document.documentElement&&[].concat(...document.body.children).forEach(i=>s.off(i,"mouseover",Le)),this._activeTrigger[ls]=!1,this._activeTrigger[Ze]=!1,this._activeTrigger[Ft]=!1,this.tip.classList.contains(ye)){const i=z(t);s.one(t,"transitionend",e),N(t,i)}else e();this._hoverState=""}}update(){this._popper!==null&&this._popper.update()}isWithContent(){return Boolean(this.getTitle())}getTipElement(){if(this.tip)return this.tip;const t=document.createElement("div");return t.innerHTML=this.config.template,this.tip=t.children[0],this.tip}setContent(){const t=this.getTipElement();this.setElementContent(f.findOne(as,t),this.getTitle()),t.classList.remove(ye,Yt)}setElementContent(t,e){if(t!==null){if(typeof e=="object"&&_t(e)){e.jquery&&(e=e[0]),this.config.html?e.parentNode!==t&&(t.innerHTML="",t.appendChild(e)):t.textContent=e.textContent;return}this.config.html?(this.config.sanitize&&(e=vr(e,this.config.allowList,this.config.sanitizeFn)),t.innerHTML=e):t.textContent=e}}getTitle(){let t=this._element.getAttribute("data-bs-original-title");return t||(t=typeof this.config.title=="function"?this.config.title.call(this._element):this.config.title),t}updateAttachment(t){return t==="right"?"end":t==="left"?"start":t}_initializeOnDelegatedTarget(t,e){const r=this.constructor.DATA_KEY;return e=e||y.get(t.delegateTarget,r),e||(e=new this.constructor(t.delegateTarget,this._getDelegateConfig()),y.set(t.delegateTarget,r,e)),e}_getOffset(){const{offset:t}=this.config;return typeof t=="string"?t.split(",").map(e=>Number.parseInt(e,10)):typeof t=="function"?e=>t(e,this._element):t}_getPopperConfig(t){const e={placement:t,modifiers:[{name:"flip",options:{fallbackPlacements:this.config.fallbackPlacements}},{name:"offset",options:{offset:this._getOffset()}},{name:"preventOverflow",options:{boundary:this.config.boundary}},{name:"arrow",options:{element:`.${this.constructor.NAME}-arrow`}},{name:"onChange",enabled:!0,phase:"afterWrite",fn:r=>this._handlePopperPlacementChange(r)}],onFirstUpdate:r=>{r.options.placement!==r.placement&&this._handlePopperPlacementChange(r)}};return{...e,...typeof this.config.popperConfig=="function"?this.config.popperConfig(e):this.config.popperConfig}}_addAttachmentClass(t){this.getTipElement().classList.add(`${Ni}-${this.updateAttachment(t)}`)}_getContainer(){return this.config.container===!1?document.body:_t(this.config.container)?this.config.container:f.findOne(this.config.container)}_getAttachment(t){return rs[t.toUpperCase()]}_setListeners(){this.config.trigger.split(" ").forEach(e=>{if(e==="click")s.on(this._element,this.constructor.Event.CLICK,this.config.selector,r=>this.toggle(r));else if(e!==ss){const r=e===Ft?this.constructor.Event.MOUSEENTER:this.constructor.Event.FOCUSIN,i=e===Ft?this.constructor.Event.MOUSELEAVE:this.constructor.Event.FOCUSOUT;s.on(this._element,r,this.config.selector,n=>this._enter(n)),s.on(this._element,i,this.config.selector,n=>this._leave(n))}}),this._hideModalHandler=()=>{this._element&&this.hide()},s.on(this._element.closest(`.${wr}`),"hide.bs.modal",this._hideModalHandler),this.config.selector?this.config={...this.config,trigger:"manual",selector:""}:this._fixTitle()}_fixTitle(){const t=this._element.getAttribute("title"),e=typeof this._element.getAttribute("data-bs-original-title");(t||e!=="string")&&(this._element.setAttribute("data-bs-original-title",t||""),t&&!this._element.getAttribute("aria-label")&&!this._element.textContent&&this._element.setAttribute("aria-label",t),this._element.setAttribute("title",""))}_enter(t,e){if(e=this._initializeOnDelegatedTarget(t,e),t&&(e._activeTrigger[t.type==="focusin"?Ze:Ft]=!0),e.getTipElement().classList.contains(Yt)||e._hoverState===Ut){e._hoverState=Ut;return}if(clearTimeout(e._timeout),e._hoverState=Ut,!e.config.delay||!e.config.delay.show){e.show();return}e._timeout=setTimeout(()=>{e._hoverState===Ut&&e.show()},e.config.delay.show)}_leave(t,e){if(e=this._initializeOnDelegatedTarget(t,e),t&&(e._activeTrigger[t.type==="focusout"?Ze:Ft]=e._element.contains(t.relatedTarget)),!e._isWithActiveTrigger()){if(clearTimeout(e._timeout),e._hoverState=Je,!e.config.delay||!e.config.delay.hide){e.hide();return}e._timeout=setTimeout(()=>{e._hoverState===Je&&e.hide()},e.config.delay.hide)}}_isWithActiveTrigger(){for(const t in this._activeTrigger)if(this._activeTrigger[t])return!0;return!1}_getConfig(t){const e=O.getDataAttributes(this._element);return Object.keys(e).forEach(r=>{es.has(r)&&delete e[r]}),t&&typeof t.container=="object"&&t.container.jquery&&(t.container=t.container[0]),t={...this.constructor.Default,...e,...typeof t=="object"&&t?t:{}},typeof t.delay=="number"&&(t.delay={show:t.delay,hide:t.delay}),typeof t.title=="number"&&(t.title=t.title.toString()),typeof t.content=="number"&&(t.content=t.content.toString()),ot(so,t,this.constructor.DefaultType),t.sanitize&&(t.template=vr(t.template,t.allowList,t.sanitizeFn)),t}_getDelegateConfig(){const t={};if(this.config)for(const e in this.config)this.constructor.Default[e]!==this.config[e]&&(t[e]=this.config[e]);return t}_cleanTipClass(){const t=this.getTipElement(),e=t.getAttribute("class").match(ts);e!==null&&e.length>0&&e.map(r=>r.trim()).forEach(r=>t.classList.remove(r))}_handlePopperPlacementChange(t){const{state:e}=t;!e||(this.tip=e.elements.popper,this._cleanTipClass(),this._addAttachmentClass(this._getAttachment(e.placement)))}static jQueryInterface(t){return this.each(function(){let e=y.get(this,mo);const r=typeof t=="object"&&t;if(!(!e&&/dispose|hide/.test(t))&&(e||(e=new Pt(this,r)),typeof t=="string")){if(typeof e[t]=="undefined")throw new TypeError(`No method named "${t}"`);e[t]()}})}}Y(so,Pt);const $i="popover",Se="bs.popover",X=`.${Se}`,Ii="bs-popover",ds=new RegExp(`(^|\\s)${Ii}\\S+`,"g"),ms={...Pt.Default,placement:"right",offset:[0,8],trigger:"click",content:"",template:'<div class="popover" role="tooltip"><div class="popover-arrow"></div><h3 class="popover-header"></h3><div class="popover-body"></div></div>'},cs={...Pt.DefaultType,content:"(string|element|function)"},ps={HIDE:`hide${X}`,HIDDEN:`hidden${X}`,SHOW:`show${X}`,SHOWN:`shown${X}`,INSERTED:`inserted${X}`,CLICK:`click${X}`,FOCUSIN:`focusin${X}`,FOCUSOUT:`focusout${X}`,MOUSEENTER:`mouseenter${X}`,MOUSELEAVE:`mouseleave${X}`},gs="fade",fs="show",us=".popover-header",bs=".popover-body";class No extends Pt{static get Default(){return ms}static get NAME(){return $i}static get DATA_KEY(){return Se}static get Event(){return ps}static get EVENT_KEY(){return X}static get DefaultType(){return cs}isWithContent(){return this.getTitle()||this._getContent()}setContent(){const t=this.getTipElement();this.setElementContent(f.findOne(us,t),this.getTitle());let e=this._getContent();typeof e=="function"&&(e=e.call(this._element)),this.setElementContent(f.findOne(bs,t),e),t.classList.remove(gs,fs)}_addAttachmentClass(t){this.getTipElement().classList.add(`${Ii}-${this.updateAttachment(t)}`)}_getContent(){return this._element.getAttribute("data-bs-content")||this.config.content}_cleanTipClass(){const t=this.getTipElement(),e=t.getAttribute("class").match(ds);e!==null&&e.length>0&&e.map(r=>r.trim()).forEach(r=>t.classList.remove(r))}static jQueryInterface(t){return this.each(function(){let e=y.get(this,Se);const r=typeof t=="object"?t:null;if(!(!e&&/dispose|hide/.test(t))&&(e||(e=new No(this,r),y.set(this,Se,e)),typeof t=="string")){if(typeof e[t]=="undefined")throw new TypeError(`No method named "${t}"`);e[t]()}})}}Y($i,No);const co="scrollspy",zi="bs.scrollspy",je=`.${zi}`,hs=".data-api",yr={offset:10,method:"auto",target:""},xs={offset:"number",method:"string",target:"(string|element)"},vs=`activate${je}`,ws=`scroll${je}`,ys=`load${je}${hs}`,_r="dropdown-item",xt="active",_s='[data-bs-spy="scroll"]',ks=".nav, .list-group",to=".nav-link",Es=".nav-item",kr=".list-group-item",Ts=".dropdown",As=".dropdown-toggle",Ss="offset",Er="position";class ie extends Q{constructor(t,e){super(t),this._scrollElement=this._element.tagName==="BODY"?window:this._element,this._config=this._getConfig(e),this._selector=`${this._config.target} ${to}, ${this._config.target} ${kr}, ${this._config.target} .${_r}`,this._offsets=[],this._targets=[],this._activeTarget=null,this._scrollHeight=0,s.on(this._scrollElement,ws,()=>this._process()),this.refresh(),this._process()}static get Default(){return yr}static get DATA_KEY(){return zi}refresh(){const t=this._scrollElement===this._scrollElement.window?Ss:Er,e=this._config.method==="auto"?t:this._config.method,r=e===Er?this._getScrollTop():0;this._offsets=[],this._targets=[],this._scrollHeight=this._getScrollHeight(),f.find(this._selector).map(n=>{const a=Oo(n),l=a?f.findOne(a):null;if(l){const d=l.getBoundingClientRect();if(d.width||d.height)return[O[e](l).top+r,a]}return null}).filter(n=>n).sort((n,a)=>n[0]-a[0]).forEach(n=>{this._offsets.push(n[0]),this._targets.push(n[1])})}dispose(){super.dispose(),s.off(this._scrollElement,je),this._scrollElement=null,this._config=null,this._selector=null,this._offsets=null,this._targets=null,this._activeTarget=null,this._scrollHeight=null}_getConfig(t){if(t={...yr,...O.getDataAttributes(this._element),...typeof t=="object"&&t?t:{}},typeof t.target!="string"&&_t(t.target)){let{id:e}=t.target;e||(e=oi(co),t.target.id=e),t.target=`#${e}`}return ot(co,t,xs),t}_getScrollTop(){return this._scrollElement===window?this._scrollElement.pageYOffset:this._scrollElement.scrollTop}_getScrollHeight(){return this._scrollElement.scrollHeight||Math.max(document.body.scrollHeight,document.documentElement.scrollHeight)}_getOffsetHeight(){return this._scrollElement===window?window.innerHeight:this._scrollElement.getBoundingClientRect().height}_process(){const t=this._getScrollTop()+this._config.offset,e=this._getScrollHeight(),r=this._config.offset+e-this._getOffsetHeight();if(this._scrollHeight!==e&&this.refresh(),t>=r){const i=this._targets[this._targets.length-1];this._activeTarget!==i&&this._activate(i);return}if(this._activeTarget&&t<this._offsets[0]&&this._offsets[0]>0){this._activeTarget=null,this._clear();return}for(let i=this._offsets.length;i--;)this._activeTarget!==this._targets[i]&&t>=this._offsets[i]&&(typeof this._offsets[i+1]=="undefined"||t<this._offsets[i+1])&&this._activate(this._targets[i])}_activate(t){this._activeTarget=t,this._clear();const e=this._selector.split(",").map(i=>`${i}[data-bs-target="${t}"],${i}[href="${t}"]`),r=f.findOne(e.join(","));r.classList.contains(_r)?(f.findOne(As,r.closest(Ts)).classList.add(xt),r.classList.add(xt)):(r.classList.add(xt),f.parents(r,ks).forEach(i=>{f.prev(i,`${to}, ${kr}`).forEach(n=>n.classList.add(xt)),f.prev(i,Es).forEach(n=>{f.children(n,to).forEach(a=>a.classList.add(xt))})})),s.trigger(this._scrollElement,vs,{relatedTarget:t})}_clear(){f.find(this._selector).filter(t=>t.classList.contains(xt)).forEach(t=>t.classList.remove(xt))}static jQueryInterface(t){return this.each(function(){const e=ie.getInstance(this)||new ie(this,typeof t=="object"?t:{});if(typeof t=="string"){if(typeof e[t]=="undefined")throw new TypeError(`No method named "${t}"`);e[t]()}})}}s.on(window,ys,()=>{f.find(_s).forEach(o=>new ie(o))});Y(co,ie);const Os="tab",De="bs.tab",le=`.${De}`,Ls=".data-api",Cs=`hide${le}`,Ds=`hidden${le}`,Ns=`show${le}`,$s=`shown${le}`,Is=`click${le}${Ls}`,zs="dropdown-menu",Gt="active",Tr="fade",Ar="show",Ms=".dropdown",Rs=".nav, .list-group",Sr=".active",Or=":scope > li > .active",Ps='[data-bs-toggle="tab"], [data-bs-toggle="pill"], [data-bs-toggle="list"]',js=".dropdown-toggle",Vs=":scope > .dropdown-menu .active";class Ve extends Q{static get DATA_KEY(){return De}show(){if(this._element.parentNode&&this._element.parentNode.nodeType===Node.ELEMENT_NODE&&this._element.classList.contains(Gt))return;let t;const e=lt(this._element),r=this._element.closest(Rs);if(r){const l=r.nodeName==="UL"||r.nodeName==="OL"?Or:Sr;t=f.find(l,r),t=t[t.length-1]}const i=t?s.trigger(t,Cs,{relatedTarget:this._element}):null;if(s.trigger(this._element,Ns,{relatedTarget:t}).defaultPrevented||i!==null&&i.defaultPrevented)return;this._activate(this._element,r);const a=()=>{s.trigger(t,Ds,{relatedTarget:this._element}),s.trigger(this._element,$s,{relatedTarget:t})};e?this._activate(e,e.parentNode,a):a()}_activate(t,e,r){const n=(e&&(e.nodeName==="UL"||e.nodeName==="OL")?f.find(Or,e):f.children(e,Sr))[0],a=r&&n&&n.classList.contains(Tr),l=()=>this._transitionComplete(t,n,r);if(n&&a){const d=z(n);n.classList.remove(Ar),s.one(n,"transitionend",l),N(n,d)}else l()}_transitionComplete(t,e,r){if(e){e.classList.remove(Gt);const n=f.findOne(Vs,e.parentNode);n&&n.classList.remove(Gt),e.getAttribute("role")==="tab"&&e.setAttribute("aria-selected",!1)}t.classList.add(Gt),t.getAttribute("role")==="tab"&&t.setAttribute("aria-selected",!0),Rt(t),t.classList.contains(Tr)&&t.classList.add(Ar);let i=t.parentNode;if(i&&i.nodeName==="LI"&&(i=i.parentNode),i&&i.classList.contains(zs)){const n=t.closest(Ms);n&&f.find(js,n).forEach(a=>a.classList.add(Gt)),t.setAttribute("aria-expanded",!0)}r&&r()}static jQueryInterface(t){return this.each(function(){const e=y.get(this,De)||new Ve(this);if(typeof t=="string"){if(typeof e[t]=="undefined")throw new TypeError(`No method named "${t}"`);e[t]()}})}}s.on(document,Is,Ps,function(o){if(["A","AREA"].includes(this.tagName)&&o.preventDefault(),At(this))return;(y.get(this,De)||new Ve(this)).show()});Y(Os,Ve);const Mi="toast",po="bs.toast",se=`.${po}`,Hs=`click.dismiss${se}`,Bs=`hide${se}`,Ws=`hidden${se}`,Ks=`show${se}`,Ys=`shown${se}`,Us="fade",Lr="hide",Xt="show",Cr="showing",Fs={animation:"boolean",autohide:"boolean",delay:"number"},Dr={animation:!0,autohide:!0,delay:5e3},Gs='[data-bs-dismiss="toast"]';class $o extends Q{constructor(t,e){super(t),this._config=this._getConfig(e),this._timeout=null,this._setListeners()}static get DefaultType(){return Fs}static get Default(){return Dr}static get DATA_KEY(){return po}show(){if(s.trigger(this._element,Ks).defaultPrevented)return;this._clearTimeout(),this._config.animation&&this._element.classList.add(Us);const e=()=>{this._element.classList.remove(Cr),this._element.classList.add(Xt),s.trigger(this._element,Ys),this._config.autohide&&(this._timeout=setTimeout(()=>{this.hide()},this._config.delay))};if(this._element.classList.remove(Lr),Rt(this._element),this._element.classList.add(Cr),this._config.animation){const r=z(this._element);s.one(this._element,"transitionend",e),N(this._element,r)}else e()}hide(){if(!this._element.classList.contains(Xt)||s.trigger(this._element,Bs).defaultPrevented)return;const e=()=>{this._element.classList.add(Lr),s.trigger(this._element,Ws)};if(this._element.classList.remove(Xt),this._config.animation){const r=z(this._element);s.one(this._element,"transitionend",e),N(this._element,r)}else e()}dispose(){this._clearTimeout(),this._element.classList.contains(Xt)&&this._element.classList.remove(Xt),super.dispose(),this._config=null}_getConfig(t){return t={...Dr,...O.getDataAttributes(this._element),...typeof t=="object"&&t?t:{}},ot(Mi,t,this.constructor.DefaultType),t}_setListeners(){s.on(this._element,Hs,Gs,()=>this.hide())}_clearTimeout(){clearTimeout(this._timeout),this._timeout=null}static jQueryInterface(t){return this.each(function(){let e=y.get(this,po);const r=typeof t=="object"&&t;if(e||(e=new $o(this,r)),typeof t=="string"){if(typeof e[t]=="undefined")throw new TypeError(`No method named "${t}"`);e[t](this)}})}}Y(Mi,$o);const Ri=document.getElementById("postForm"),Xs=Ri[1],go=Ri[2];go.setAttribute("disabled","true");Xs.addEventListener("input",o=>{if(!(o.target instanceof HTMLTextAreaElement))return;o.target.value.replace(/\r?\n/g,"").replace(/\s+/g,"")?go.removeAttribute("disabled"):go.setAttribute("disabled","true")});const qs=document.getElementsByClassName("commentInfo"),Pi=document.getElementById("commentPostModal"),Qs=document.getElementById("modalBody"),Js=Qs.children[0],Ne=document.getElementById("postContent"),$e=document.getElementById("commentInsertButton"),ji=document.getElementById("commentErrorMessage"),Vi=document.getElementById("flashSuccessMessage"),fo=Js.children[0].value,Zs=new kt(Pi);let Io=null,Qt=null;$e.setAttribute("disabled","true");Ne.addEventListener("input",o=>{Ne.value.replace(/\r|\n/g,"").replace(/\s+/g,"")!=""?$e.removeAttribute("disabled"):$e.setAttribute("disabled","true")});Pi.addEventListener("hide.bs.modal",()=>{Io=null,Ne.value="",ji.innerText=""});Array.from(qs).map(o=>o.addEventListener("click",t=>{t.target instanceof HTMLSpanElement&&(Vi.innerText="",Io=t.target.dataset.postId,Qt=t.target.nextElementSibling)}));$e.addEventListener("click",o=>{const t=Ne.value;Fi.post("/comment",{postId:Io,content:t},{headers:{"Csrf-Token":fo}}).then(e=>{Vi.innerText="\u6295\u7A3F\u5B8C\u4E86\u3057\u307E\u3057\u305F",Qt.innerText==""?Qt.innerText=1:Qt.innerText=parseInt(Qt.innerText)+1,Zs.hide()}).catch(e=>{e.response.status==401&&(location.href="/users/sign-in?=needSignIn");let r=e.response.data.content[0];ji.innerText=r})});const td=document.getElementsByClassName("likeInfo");Array.from(td).map(o=>{const t=o.children[0],e=o.children[1],r=n=>{e.innerText=n.data,t.classList.remove("like-btn"),t.classList.add("unlike-btn")},i=n=>{n.data==0?e.innerText="":e.innerText=n.data,t.classList.remove("unlike-btn"),t.classList.add("like-btn")};t.addEventListener("click",n=>{if(!(n.target instanceof HTMLElement))return;const a=n.target.dataset.postId;n.target.classList.contains("like-btn")?Gi(a,fo,r):Xi(a,fo,i)})});
